@@ -3,9 +3,10 @@ package alter
 import (
 	"strconv"
 
+	"github.com/deviceinsight/kafkactl/v5/internal"
+
 	"github.com/deviceinsight/kafkactl/v5/cmd/validation"
 	"github.com/deviceinsight/kafkactl/v5/internal/k8s"
-	"github.com/deviceinsight/kafkactl/v5/internal/output"
 	"github.com/deviceinsight/kafkactl/v5/internal/partition"
 	"github.com/deviceinsight/kafkactl/v5/internal/topic"
 	"github.com/pkg/errors"
@@ -20,21 +21,21 @@ func newAlterPartitionCmd() *cobra.Command {
 		Use:   "partition TOPIC PARTITION",
 		Short: "alter a partition",
 		Args:  cobra.ExactArgs(2),
-		Run: func(cmd *cobra.Command, args []string) {
-			if !k8s.NewOperation().TryRun(cmd, args) {
-
-				var partitionID int32
-
-				if i, err := strconv.ParseInt(args[1], 10, 64); err != nil {
-					output.Fail(errors.Errorf("argument 2 needs to be a partition %s", args[1]))
-				} else {
-					partitionID = int32(i)
-				}
-
-				if err := (&partition.Operation{}).AlterPartition(args[0], partitionID, flags); err != nil {
-					output.Fail(err)
-				}
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if internal.IsKubernetesEnabled() {
+				return k8s.NewOperation().Run(cmd, args)
 			}
+
+			var (
+				partitionID int64
+				err         error
+			)
+
+			if partitionID, err = strconv.ParseInt(args[1], 10, 64); err != nil {
+				return errors.Errorf("argument 2 needs to be a partition %s", args[1])
+			}
+
+			return (&partition.Operation{}).AlterPartition(args[0], int32(partitionID), flags)
 		},
 		PreRunE: func(cmd *cobra.Command, _ []string) error {
 			return validation.ValidateAtLeastOneRequiredFlag(cmd)
