@@ -9,10 +9,12 @@ func NewAlterCmd() *cobra.Command {
 	var cmdAlter = &cobra.Command{
 		Use:     "alter",
 		Aliases: []string{"edit"},
-		Short:   "alter topics, partitions",
+		Short:   "alter topics, partitions, brokers, users",
 	}
 
 	cmdAlter.AddCommand(newAlterTopicCmd())
 	cmdAlter.AddCommand(newAlterPartitionCmd())
+	cmdAlter.AddCommand(newAlterBrokerCmd())
+	cmdAlter.AddCommand(newAlterUserCmd())
 	return cmdAlter
 }
