@@ -1,10 +1,10 @@
 package config
 
 import (
-	"os"
+	"bytes"
+	"fmt"
 
 	"github.com/deviceinsight/kafkactl/v5/internal/output"
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -13,16 +13,18 @@ func newViewCmd() *cobra.Command {
 
 	var cmdView = &cobra.Command{
 		Use:   "view",
-		Short: "show contents of config file",
-		Long:  `Shows the contents of the config file that is currently used`,
-		Run: func(_ *cobra.Command, _ []string) {
+		Short: "show current config",
+		Long:  `Shows the merged config that is currently used`,
+		RunE: func(_ *cobra.Command, _ []string) error {
 
-			yamlFile, err := os.ReadFile(viper.ConfigFileUsed())
-			if err != nil {
-				output.Fail(errors.Wrap(err, "unable to read config"))
+			c := new(bytes.Buffer)
+
+			if err := viper.WriteConfigTo(c); err != nil {
+				return fmt.Errorf("unable to write config: %v", err)
 			}
 
-			output.Infof("%s", yamlFile)
+			output.Infof("%s", c.String())
+			return nil
 		},
 	}
 
