@@ -2,21 +2,26 @@ package consume_test
 
 import (
 	"encoding/hex"
+	"fmt"
 	"path/filepath"
 	"strconv"
 	"strings"
 	"testing"
 	"time"
 
+	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/reflect/protoreflect"
+	"google.golang.org/protobuf/types/dynamicpb"
+
+	"github.com/deviceinsight/kafkactl/v5/internal"
 	"github.com/deviceinsight/kafkactl/v5/internal/helpers/protobuf"
+	"github.com/riferrei/srclient"
 
 	"github.com/deviceinsight/kafkactl/v5/internal/testutil"
-	"github.com/jhump/protoreflect/dynamic"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 func TestConsumeWithKeyAndValueIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic")
@@ -33,7 +38,6 @@ func TestConsumeWithKeyAndValueIntegration(t *testing.T) {
 }
 
 func TestConsumeWithPartitionAndValueIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic", "--partitions", "2")
@@ -50,7 +54,6 @@ func TestConsumeWithPartitionAndValueIntegration(t *testing.T) {
 }
 
 func TestConsumeWithEmptyPartitionsIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic", "--partitions", "10")
@@ -67,7 +70,6 @@ func TestConsumeWithEmptyPartitionsIntegration(t *testing.T) {
 }
 
 func TestConsumeTailIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic", "--partitions", "10")
@@ -94,8 +96,7 @@ func TestConsumeTailIntegration(t *testing.T) {
 	testutil.AssertEquals(t, "test-key-2#test-value-2b", messages[2])
 }
 
-func TestConsumeFromTimestamp(t *testing.T) {
-
+func TestConsumeFromTimestampIntegration(t *testing.T) {
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic", "--partitions", "2")
@@ -117,25 +118,25 @@ func TestConsumeFromTimestamp(t *testing.T) {
 	testutil.ProduceMessageOnPartition(t, topicName, "key-2", "g", 1, 3)
 	testutil.ProduceMessageOnPartition(t, topicName, "key-1", "h", 0, 3)
 
-	//test --from-timestamp with --to-timestamp with formatted dates
+	// test --from-timestamp with --to-timestamp with formatted dates
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
-	t1Formatted := time.UnixMilli(t1).Format("2006-01-02T15:04:05.123Z")
-	t2Formatted := time.UnixMilli(t2).Format("2006-01-02T15:04:05.123Z")
+	t1Formatted := time.UnixMilli(t1).Format("2006-01-02T15:04:05.000Z")
+	t2Formatted := time.UnixMilli(t2).Format("2006-01-02T15:04:05.000Z")
 	if _, err := kafkaCtl.Execute("consume", topicName, "--from-timestamp", t1Formatted, "--to-timestamp", t2Formatted); err != nil {
 		t.Fatalf("failed to execute command: %v", err)
 	}
 	messages := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
 	testutil.AssertArraysEquals(t, []string{"c", "d", "e", "f"}, messages)
 
-	//test --from-timestamp with --to-timestamp with unix epoch millis timestamps
+	// test --from-timestamp with --to-timestamp with unix epoch millis timestamps
 	kafkaCtl = testutil.CreateKafkaCtlCommand()
-	if _, err := kafkaCtl.Execute("consume", topicName, "--from-timestamp", strconv.FormatInt(t2, 10), "--to-timestamp", strconv.FormatInt(t2, 10)); err != nil {
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-timestamp", strconv.FormatInt(t1, 10), "--to-timestamp", strconv.FormatInt(t2, 10)); err != nil {
 		t.Fatalf("failed to execute command: %v", err)
 	}
 	messages = strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
 	testutil.AssertArraysEquals(t, []string{"c", "d", "e", "f"}, messages)
 
-	//test --from-timestamp with --max-messages (--partitions present for reproducibility)
+	// test --from-timestamp with --max-messages (--partitions present for reproducibility)
 	kafkaCtl = testutil.CreateKafkaCtlCommand()
 	if _, err := kafkaCtl.Execute("consume", topicName, "--from-timestamp", strconv.FormatInt(t1, 10), "--max-messages", strconv.Itoa(2), "--partitions", strconv.Itoa(1)); err != nil {
 		t.Fatalf("failed to execute command: %v", err)
@@ -143,7 +144,7 @@ func TestConsumeFromTimestamp(t *testing.T) {
 	messages = strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
 	testutil.AssertArraysEquals(t, []string{"c", "d"}, messages)
 
-	//test --from-timestamp with --exit
+	// test --from-timestamp with --exit
 	kafkaCtl = testutil.CreateKafkaCtlCommand()
 	if _, err := kafkaCtl.Execute("consume", topicName, "--from-timestamp", strconv.FormatInt(t2, 10), "--exit"); err != nil {
 		t.Fatalf("failed to execute command: %v", err)
@@ -152,8 +153,215 @@ func TestConsumeFromTimestamp(t *testing.T) {
 	testutil.AssertArraysEquals(t, []string{"g", "h"}, messages)
 }
 
-func TestConsumeToTimestamp(t *testing.T) {
+func TestConsumeRegistryProtobufWithNestedDependenciesIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
 
+	bazMsg := `syntax = "proto3";
+  package baz;
+
+  message Baz {
+    string field = 1;
+  }
+  `
+
+	barMsg := `syntax = "proto3";
+  package bar;
+
+  import "baz/protobuf/baz.proto";
+
+  message Bar {
+    baz.Baz bazField = 1;
+  }
+  `
+
+	fooMsg := `syntax = "proto3";
+  package foo;
+
+  import "bar/protobuf/bar.proto";
+
+  message Foo {
+    bar.Bar barField = 1;
+  }`
+
+	value := `{"barField":{"bazField":{"field":"value"}}}`
+
+	testutil.RegisterSchema(t, "baz", bazMsg, srclient.Protobuf)
+	testutil.RegisterSchema(t, "bar", barMsg, srclient.Protobuf, srclient.Reference{Name: "baz/protobuf/baz.proto", Version: 1, Subject: "baz"})
+	topicName := testutil.CreateTopic(t, "consume-topic")
+	testutil.RegisterSchema(t, topicName+"-value", fooMsg, srclient.Protobuf, srclient.Reference{Name: "bar/protobuf/bar.proto", Version: 1, Subject: "bar"})
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "test-key", "--value", value); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--print-keys"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, fmt.Sprintf("test-key#%s", value), kafkaCtl.GetStdOut())
+}
+
+func TestConsumeRegistryEmitDefaultValuesIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	singleMsg := `syntax = "proto3";
+	
+		message MyMessage {
+		  message Metadata {
+			string key = 1;
+			string value = 2;
+		  }
+		  bool active = 3;
+		  repeated Metadata metadata = 4;
+		}
+		`
+
+	listMsg := `syntax = "proto3";
+
+import "my-message.proto";
+
+message MyList {
+  repeated MyMessage messages = 3;
+}
+  `
+
+	value := `{
+  "messages": [
+    {
+      "active": true,
+      "metadata": []
+    },
+    {
+      "active": false,
+      "metadata": []
+    }
+  ]
+}`
+
+	topicName := testutil.CreateTopic(t, "consume-topic")
+	testutil.RegisterSchema(t, "my-message", singleMsg, srclient.Protobuf)
+	testutil.RegisterSchema(t, topicName+"-value", listMsg, srclient.Protobuf, srclient.Reference{Name: "my-message.proto", Version: 1, Subject: "my-message"})
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "test-key", "--value", value); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--print-keys", "--proto-marshal-option", "emitdefaultvalues=true"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, fmt.Sprintf("test-key#%s", removeWhitespace(value)), removeWhitespace(kafkaCtl.GetStdOut()))
+}
+
+func TestConsumeRegistryProtobufWithWellKnowTypeIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	newFooMsg := `syntax = "proto3";
+  package new.foo;
+
+  import "google/protobuf/timestamp.proto";
+
+  message Foo {
+    google.protobuf.Timestamp field = 1;
+  }
+  `
+	value := `{"field":"2025-06-07T11:11:11Z"}`
+
+	testutil.RegisterSchema(t, "new.foo", newFooMsg, srclient.Protobuf)
+	topicName := testutil.CreateTopic(t, "consume-topic")
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "test-key", "--value", value); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--print-keys"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, fmt.Sprintf("test-key#%s", value), kafkaCtl.GetStdOut())
+}
+
+func TestConsumeRegistryProtobufWithNestedProtoIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	// example from https://docs.confluent.io/platform/current/schema-registry/fundamentals/serdes-develop/index.html#wire-format
+	schema := `syntax = "proto3";
+		package test.package;
+		
+		message MessageA {
+			message MessageB {
+				message MessageC {
+					string fieldC = 1;
+				}
+			}
+			message MessageD {
+				string fieldD = 1;
+			}
+			message MessageE {
+				message MessageF {
+					string fieldF = 1;
+				}
+				message MessageG {
+					string fieldG = 1;
+				}
+			}
+		}
+		message MessageH {
+			message MessageI {
+				string fieldI = 1;
+			}
+			MessageI fieldH = 1;
+		}`
+
+	testCases := []struct {
+		name      string
+		value     string
+		protoType string
+	}{
+		{
+			name:      "test_MessageH",
+			value:     `{"fieldH":{"fieldI":"value"}}`,
+			protoType: "test.package.MessageH",
+		},
+		{
+			name:      "test_MessageI",
+			value:     `{"fieldI":"value"}`,
+			protoType: "test.package.MessageH.MessageI",
+		},
+		{
+			name:      "test_MessageG",
+			value:     `{"fieldG":"value"}`,
+			protoType: "test.package.MessageA.MessageE.MessageG",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			topicName := testutil.CreateTopic(t, "consume-topic")
+			testutil.RegisterSchema(t, topicName+"-value", schema, srclient.Protobuf)
+
+			kafkaCtl := testutil.CreateKafkaCtlCommand()
+			if _, err := kafkaCtl.Execute("produce", topicName, "--value-proto-type", tc.protoType, "--value", tc.value); err != nil {
+				t.Fatalf("failed to execute command: %v", err)
+			}
+			testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+			if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit"); err != nil {
+				t.Fatalf("failed to execute command: %v", err)
+			}
+
+			testutil.AssertEquals(t, tc.value, kafkaCtl.GetStdOut())
+		})
+	}
+}
+
+func TestConsumeToTimestampIntegration(t *testing.T) {
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic", "--partitions", "2")
@@ -175,7 +383,7 @@ func TestConsumeToTimestamp(t *testing.T) {
 	testutil.ProduceMessageOnPartition(t, topicName, "key-2", "g", 1, 3)
 	testutil.ProduceMessageOnPartition(t, topicName, "key-1", "h", 0, 3)
 
-	//test --from-beginning with --to-timestamp
+	// test --from-beginning with --to-timestamp
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
 	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--to-timestamp", strconv.FormatInt(t1, 10)); err != nil {
 		t.Fatalf("failed to execute command: %v", err)
@@ -183,7 +391,7 @@ func TestConsumeToTimestamp(t *testing.T) {
 	messages := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
 	testutil.AssertArraysEquals(t, []string{"a", "b"}, messages)
 
-	//test --to-timestamp with --tail
+	// test --to-timestamp with --tail
 	kafkaCtl = testutil.CreateKafkaCtlCommand()
 	if _, err := kafkaCtl.Execute("consume", topicName, "--to-timestamp", strconv.FormatInt(t2, 10), "--tail", strconv.Itoa(4)); err != nil {
 		t.Fatalf("failed to execute command: %v", err)
@@ -193,7 +401,6 @@ func TestConsumeToTimestamp(t *testing.T) {
 }
 
 func TestConsumeWithKeyAndValueAsBase64Integration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic")
@@ -213,7 +420,6 @@ func TestConsumeWithKeyAndValueAsBase64Integration(t *testing.T) {
 }
 
 func TestConsumeWithKeyAndValueAsHexIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic")
@@ -233,7 +439,6 @@ func TestConsumeWithKeyAndValueAsHexIntegration(t *testing.T) {
 }
 
 func TestConsumeWithKeyAndValueAutoDetectBinaryValueIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "consume-topic")
@@ -259,7 +464,6 @@ func TestConsumeWithKeyAndValueAutoDetectBinaryValueIntegration(t *testing.T) {
 }
 
 func TestAvroDeserializationErrorHandlingIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	valueSchema := `{
@@ -275,7 +479,7 @@ func TestAvroDeserializationErrorHandlingIntegration(t *testing.T) {
 	value := `{"name":"Peter Mueller"}`
 	value2 := `{"name":"Peter Pan"}`
 
-	topicName := testutil.CreateAvroTopic(t, "avro-topic", "", valueSchema)
+	topicName := testutil.CreateTopicWithSchema(t, "avro-topic", "", valueSchema, srclient.Avro)
 
 	group := testutil.CreateConsumerGroup(t, "avro-topic-consumer-group", topicName)
 
@@ -307,19 +511,23 @@ func TestAvroDeserializationErrorHandlingIntegration(t *testing.T) {
 	testutil.AssertEquals(t, "message produced (partition=0\toffset=2)", kafkaCtl.GetStdOut())
 
 	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit"); err != nil {
-		testutil.AssertErrorContains(t, "failed to find avro schema", err)
-		testutil.AssertEquals(t, value, kafkaCtl.GetStdOut())
+		t.Fatalf("failed to execute command: %v", err)
 	} else {
-		t.Fatalf("expected consumer to fail")
+		results := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+		testutil.AssertContains(t, value, results)
+		testutil.AssertContains(t, "no-avro", results)
+		testutil.AssertContains(t, value2, results)
 	}
 
 	kafkaCtl = testutil.CreateKafkaCtlCommand()
 
 	if _, err := kafkaCtl.Execute("consume", topicName, "--group", group, "--max-messages", "3"); err != nil {
-		testutil.AssertErrorContains(t, "failed to find avro schema", err)
-		testutil.AssertEquals(t, value, kafkaCtl.GetStdOut())
+		t.Fatalf("failed to execute command: %v", err)
 	} else {
-		t.Fatalf("expected consumer to fail")
+		results := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+		testutil.AssertContains(t, value, results)
+		testutil.AssertContains(t, "no-avro", results)
+		testutil.AssertContains(t, value2, results)
 	}
 }
 
@@ -332,15 +540,15 @@ func TestProtobufConsumeProtoFileIntegration(t *testing.T) {
 
 	protoPath := filepath.Join(testutil.RootDir, "internal", "testutil", "testdata")
 	now := time.Date(2021, time.December, 1, 14, 10, 12, 0, time.UTC)
-	pbMessageDesc := protobuf.ResolveMessageType(protobuf.SearchContext{
+	pbMessageDesc := protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtoImportPaths: []string{protoPath},
 		ProtoFiles:       []string{"msg.proto"},
 	}, "TopicMessage")
-	pbMessage := dynamic.NewMessage(pbMessageDesc)
-	pbMessage.SetFieldByNumber(1, timestamppb.New(now))
-	pbMessage.SetFieldByNumber(2, int64(1))
+	pbMessage := dynamicpb.NewMessage(pbMessageDesc)
+	pbMessage.Set(pbMessageDesc.Fields().Get(0), protoreflect.ValueOfMessage(timestamppb.New(now).ProtoReflect()))
+	pbMessage.Set(pbMessageDesc.Fields().Get(1), protoreflect.ValueOfInt64(1))
 
-	value, err := pbMessage.Marshal()
+	value, err := proto.Marshal(pbMessage)
 	if err != nil {
 		t.Fatalf("Failed to marshal proto message: %s", err)
 	}
@@ -356,7 +564,10 @@ func TestProtobufConsumeProtoFileIntegration(t *testing.T) {
 		t.Fatalf("failed to execute command: %v", err)
 	}
 
-	testutil.AssertEquals(t, `{"producedAt":"2021-12-01T14:10:12Z","num":"1"}`, kafkaCtl.GetStdOut())
+	// https://github.com/golang/protobuf/issues/1082
+	output := strings.ReplaceAll(kafkaCtl.GetStdOut(), " ", "")
+
+	testutil.AssertEquals(t, `{"producedAt":"2021-12-01T14:10:12Z","num":"1"}`, output)
 }
 
 func TestProtobufConsumeProtoFileWithoutProtoImportPathIntegration(t *testing.T) {
@@ -368,15 +579,15 @@ func TestProtobufConsumeProtoFileWithoutProtoImportPathIntegration(t *testing.T)
 
 	protoPath := filepath.Join(testutil.RootDir, "internal", "testutil", "testdata")
 	now := time.Date(2021, time.December, 1, 14, 10, 12, 0, time.UTC)
-	pbMessageDesc := protobuf.ResolveMessageType(protobuf.SearchContext{
+	pbMessageDesc := protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtoImportPaths: []string{protoPath},
 		ProtoFiles:       []string{"msg.proto"},
 	}, "TopicMessage")
-	pbMessage := dynamic.NewMessage(pbMessageDesc)
-	pbMessage.SetFieldByNumber(1, timestamppb.New(now))
-	pbMessage.SetFieldByNumber(2, int64(1))
+	pbMessage := dynamicpb.NewMessage(pbMessageDesc)
+	pbMessage.Set(pbMessageDesc.Fields().Get(0), protoreflect.ValueOfMessage(timestamppb.New(now).ProtoReflect()))
+	pbMessage.Set(pbMessageDesc.Fields().Get(1), protoreflect.ValueOfInt64(1))
 
-	value, err := pbMessage.Marshal()
+	value, err := proto.Marshal(pbMessage)
 	if err != nil {
 		t.Fatalf("Failed to marshal proto message: %s", err)
 	}
@@ -392,7 +603,10 @@ func TestProtobufConsumeProtoFileWithoutProtoImportPathIntegration(t *testing.T)
 		t.Fatalf("failed to execute command: %v", err)
 	}
 
-	testutil.AssertEquals(t, `{"producedAt":"2021-12-01T14:10:12Z","num":"1"}`, kafkaCtl.GetStdOut())
+	// https://github.com/golang/protobuf/issues/1082
+	output := strings.ReplaceAll(kafkaCtl.GetStdOut(), " ", "")
+
+	testutil.AssertEquals(t, `{"producedAt":"2021-12-01T14:10:12Z","num":"1"}`, output)
 }
 
 func TestConsumeTombstoneWithProtoFileIntegration(t *testing.T) {
@@ -424,14 +638,14 @@ func TestProtobufConsumeProtosetFileIntegration(t *testing.T) {
 
 	protoPath := filepath.Join(testutil.RootDir, "internal", "testutil", "testdata", "msg.protoset")
 	now := time.Date(2021, time.December, 1, 14, 10, 12, 0, time.UTC)
-	pbMessageDesc := protobuf.ResolveMessageType(protobuf.SearchContext{
+	pbMessageDesc := protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtosetFiles: []string{protoPath},
 	}, "TopicMessage")
-	pbMessage := dynamic.NewMessage(pbMessageDesc)
-	pbMessage.SetFieldByNumber(1, timestamppb.New(now))
-	pbMessage.SetFieldByNumber(2, int64(1))
+	pbMessage := dynamicpb.NewMessage(pbMessageDesc)
+	pbMessage.Set(pbMessageDesc.Fields().Get(0), protoreflect.ValueOfMessage(timestamppb.New(now).ProtoReflect()))
+	pbMessage.Set(pbMessageDesc.Fields().Get(1), protoreflect.ValueOfInt64(1))
 
-	value, err := pbMessage.Marshal()
+	value, err := proto.Marshal(pbMessage)
 	if err != nil {
 		t.Fatalf("Failed to marshal proto message: %s", err)
 	}
@@ -447,7 +661,10 @@ func TestProtobufConsumeProtosetFileIntegration(t *testing.T) {
 		t.Fatalf("failed to execute command: %v", err)
 	}
 
-	testutil.AssertEquals(t, `{"producedAt":"2021-12-01T14:10:12Z","num":"1"}`, kafkaCtl.GetStdOut())
+	// https://github.com/golang/protobuf/issues/1082
+	output := strings.ReplaceAll(kafkaCtl.GetStdOut(), " ", "")
+
+	testutil.AssertEquals(t, `{"producedAt":"2021-12-01T14:10:12Z","num":"1"}`, output)
 }
 
 func TestProtobufConsumeProtoFileErrNoMessageIntegration(t *testing.T) {
@@ -458,6 +675,25 @@ func TestProtobufConsumeProtoFileErrNoMessageIntegration(t *testing.T) {
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
 
 	protoPath := filepath.Join(testutil.RootDir, "internal", "testutil", "testdata", "msg.protoset")
+	now := time.Date(2021, time.December, 1, 14, 10, 12, 0, time.UTC)
+	pbMessageDesc := protobuf.ResolveMessageType(internal.ProtobufConfig{
+		ProtosetFiles: []string{protoPath},
+	}, "TopicMessage")
+	pbMessage := dynamicpb.NewMessage(pbMessageDesc)
+	pbMessage.Set(pbMessageDesc.Fields().Get(0), protoreflect.ValueOfMessage(timestamppb.New(now).ProtoReflect()))
+	pbMessage.Set(pbMessageDesc.Fields().Get(1), protoreflect.ValueOfInt64(1))
+
+	value, err := proto.Marshal(pbMessage)
+	if err != nil {
+		t.Fatalf("Failed to marshal proto message: %s", err)
+	}
+
+	// produce valid pb message
+	if _, err := kafkaCtl.Execute("produce", pbTopic, "--key", "test-key", "--value", hex.EncodeToString(value), "--value-encoding", "hex"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
 
 	if _, err := kafkaCtl.Execute("consume", pbTopic, "--from-beginning", "--exit", "--proto-import-path", protoPath, "--proto-file", "msg.proto", "--value-proto-type", "NonExisting"); err != nil {
 		testutil.AssertErrorContains(t, "not found in provided files", err)
@@ -483,14 +719,35 @@ func TestProtobufConsumeProtoFileErrDecodeIntegration(t *testing.T) {
 	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
 
 	if _, err := kafkaCtl.Execute("consume", pbTopic, "--from-beginning", "--exit", "--proto-import-path", protoPath, "--proto-file", "msg.proto", "--value-proto-type", "TopicMessage"); err != nil {
-		testutil.AssertErrorContains(t, "value decode failed: proto: bad wiretype", err)
+		testutil.AssertErrorContains(t, "cannot parse invalid wire-format data", err)
 	} else {
 		t.Fatal("Expected consumer to fail")
 	}
 }
 
-func TestConsumeGroupIntegration(t *testing.T) {
+func TestConsumeGroupMaxMessagesDoNotOverConsumeIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+	prefix := "consume-group-max-messages-test"
+	topicName := testutil.CreateTopic(t, prefix+"topic")
+	group := testutil.CreateConsumerGroup(t, prefix+"group", topicName)
 
+	msgs := []string{"value1", "value2", "value3", "value4", "value5"}
+
+	for i, msg := range msgs {
+		testutil.ProduceMessage(t, topicName, "test-key", msg, 0, int64(i))
+	}
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+	for _, expectedMsg := range msgs {
+		if _, err := kafkaCtl.Execute("consume", topicName, "--group", group, "--max-messages", "1"); err != nil {
+			t.Fatalf("failed to execute command: %v", err)
+		}
+		results := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+		testutil.AssertContains(t, expectedMsg, results)
+	}
+}
+
+func TestConsumeGroupIntegration(t *testing.T) {
 	testutil.StartIntegrationTest(t)
 
 	prefix := "consume-group-"
@@ -524,7 +781,6 @@ func TestConsumeGroupIntegration(t *testing.T) {
 }
 
 func TestConsumeAutoCompletionIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	prefix := "consume-complete-"
@@ -548,7 +804,6 @@ func TestConsumeAutoCompletionIntegration(t *testing.T) {
 }
 
 func TestConsumeGroupCompletionIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	prefix := "consume-group-complete-"
@@ -574,7 +829,6 @@ func TestConsumeGroupCompletionIntegration(t *testing.T) {
 }
 
 func TestConsumePartitionsK8sIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTestWithContext(t, "k8s-mock")
 
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
@@ -603,7 +857,6 @@ func TestConsumePartitionsK8sIntegration(t *testing.T) {
 		},
 	} {
 		t.Run(test.description, func(t *testing.T) {
-
 			if _, err := kafkaCtl.Execute(test.args...); err != nil {
 				t.Fatalf("failed to execute command: %v", err)
 			}
@@ -614,5 +867,186 @@ func TestConsumePartitionsK8sIntegration(t *testing.T) {
 				testutil.AssertContainSubstring(t, wanted, output)
 			}
 		})
+	}
+}
+
+func removeWhitespace(s string) string {
+	s = strings.ReplaceAll(s, " ", "")
+	s = strings.ReplaceAll(s, "\t", "")
+	s = strings.ReplaceAll(s, "\n", "")
+	s = strings.ReplaceAll(s, "\v", "")
+	s = strings.ReplaceAll(s, "\f", "")
+	s = strings.ReplaceAll(s, "\r", "")
+	return s
+}
+
+func TestConsumeFilterByKeyIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	topicName := testutil.CreateTopic(t, "consume-filter-topic")
+
+	testutil.ProduceMessage(t, topicName, "user-123", "value1", 0, 0)
+	testutil.ProduceMessage(t, topicName, "admin-456", "value2", 0, 1)
+	testutil.ProduceMessage(t, topicName, "user-789", "value3", 0, 2)
+	testutil.ProduceMessage(t, topicName, "guest-111", "value4", 0, 3)
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--filter-key", "user-*", "--print-keys"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	messages := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+
+	if len(messages) != 2 {
+		t.Fatalf("expected 2 messages, got %d: %v", len(messages), messages)
+	}
+
+	testutil.AssertEquals(t, "user-123#value1", messages[0])
+	testutil.AssertEquals(t, "user-789#value3", messages[1])
+}
+
+func TestConsumeFilterByKeyAlternativesIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	topicName := testutil.CreateTopic(t, "consume-filter-topic")
+
+	testutil.ProduceMessage(t, topicName, "user-123", "value1", 0, 0)
+	testutil.ProduceMessage(t, topicName, "admin-456", "value2", 0, 1)
+	testutil.ProduceMessage(t, topicName, "guest-789", "value3", 0, 2)
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--filter-key", "{user,admin}-*", "--print-keys"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	messages := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+
+	if len(messages) != 2 {
+		t.Fatalf("expected 2 messages, got %d: %v", len(messages), messages)
+	}
+
+	testutil.AssertEquals(t, "user-123#value1", messages[0])
+	testutil.AssertEquals(t, "admin-456#value2", messages[1])
+}
+
+func TestConsumeFilterByValueIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	topicName := testutil.CreateTopic(t, "consume-filter-topic")
+
+	testutil.ProduceMessage(t, topicName, "key1", "error: something failed", 0, 0)
+	testutil.ProduceMessage(t, topicName, "key2", "success: operation completed", 0, 1)
+	testutil.ProduceMessage(t, topicName, "key3", "error: connection timeout", 0, 2)
+	testutil.ProduceMessage(t, topicName, "key4", "info: processing", 0, 3)
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--filter-value", "error:*"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	messages := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+
+	if len(messages) != 2 {
+		t.Fatalf("expected 2 messages, got %d: %v", len(messages), messages)
+	}
+
+	testutil.AssertEquals(t, "error: something failed", messages[0])
+	testutil.AssertEquals(t, "error: connection timeout", messages[1])
+}
+
+func TestConsumeFilterByHeaderIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	topicName := testutil.CreateTopic(t, "consume-filter-topic")
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "key1", "--value", "value1", "-H", "env:prod", "-H", "trace-id:abc-123"); err != nil {
+		t.Fatalf("failed to produce message: %v", err)
+	}
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "key2", "--value", "value2", "-H", "env:dev", "-H", "trace-id:xyz-456"); err != nil {
+		t.Fatalf("failed to produce message: %v", err)
+	}
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "key3", "--value", "value3", "-H", "env:prod", "-H", "trace-id:abc-789"); err != nil {
+		t.Fatalf("failed to produce message: %v", err)
+	}
+
+	kafkaCtl = testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--filter-header", "trace-id=abc-*", "--print-keys"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	messages := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+
+	if len(messages) != 2 {
+		t.Fatalf("expected 2 messages, got %d: %v", len(messages), messages)
+	}
+
+	testutil.AssertEquals(t, "key1#value1", messages[0])
+	testutil.AssertEquals(t, "key3#value3", messages[1])
+}
+
+func TestConsumeFilterANDLogicIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	topicName := testutil.CreateTopic(t, "consume-filter-topic")
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "user-123", "--value", "error occurred", "-H", "env:prod"); err != nil {
+		t.Fatalf("failed to produce message: %v", err)
+	}
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "user-456", "--value", "success", "-H", "env:prod"); err != nil {
+		t.Fatalf("failed to produce message: %v", err)
+	}
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "admin-789", "--value", "error occurred", "-H", "env:prod"); err != nil {
+		t.Fatalf("failed to produce message: %v", err)
+	}
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "user-111", "--value", "error occurred", "-H", "env:dev"); err != nil {
+		t.Fatalf("failed to produce message: %v", err)
+	}
+
+	kafkaCtl = testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--filter-key", "user-*", "--filter-value", "error*", "--filter-header", "env=prod", "--print-keys"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	messages := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+
+	if len(messages) != 1 {
+		t.Fatalf("expected 1 message, got %d: %v", len(messages), messages)
+	}
+
+	testutil.AssertEquals(t, "user-123#error occurred", messages[0])
+}
+
+func TestConsumeFilterNoMatchIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	topicName := testutil.CreateTopic(t, "consume-filter-topic")
+
+	testutil.ProduceMessage(t, topicName, "user-123", "value1", 0, 0)
+	testutil.ProduceMessage(t, topicName, "admin-456", "value2", 0, 1)
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--filter-key", "guest-*"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	output := strings.TrimSpace(kafkaCtl.GetStdOut())
+
+	if output != "" {
+		t.Fatalf("expected no output for non-matching filter, got: %q", output)
 	}
 }
