@@ -1,10 +1,10 @@
 package create
 
 import (
+	"github.com/deviceinsight/kafkactl/v5/internal"
 	"github.com/deviceinsight/kafkactl/v5/internal/acl"
 	"github.com/deviceinsight/kafkactl/v5/internal/consumergroups"
 	"github.com/deviceinsight/kafkactl/v5/internal/k8s"
-	"github.com/deviceinsight/kafkactl/v5/internal/output"
 	"github.com/deviceinsight/kafkactl/v5/internal/topic"
 	"github.com/spf13/cobra"
 )
@@ -18,12 +18,11 @@ func newCreateACLCmd() *cobra.Command {
 		Aliases: []string{"acl"},
 		Short:   "create an acl",
 		Args:    cobra.MaximumNArgs(0),
-		Run: func(cmd *cobra.Command, args []string) {
-			if !k8s.NewOperation().TryRun(cmd, args) {
-				if err := (&acl.Operation{}).CreateACL(flags); err != nil {
-					output.Fail(err)
-				}
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if internal.IsKubernetesEnabled() {
+				return k8s.NewOperation().Run(cmd, args)
 			}
+			return (&acl.Operation{}).CreateACL(flags)
 		},
 		ValidArgsFunction: acl.CompleteCreateACL,
 	}
