@@ -1,9 +1,9 @@
 package deletion
 
 import (
+	"github.com/deviceinsight/kafkactl/v5/internal"
 	"github.com/deviceinsight/kafkactl/v5/internal/consumergroups"
 	"github.com/deviceinsight/kafkactl/v5/internal/k8s"
-	"github.com/deviceinsight/kafkactl/v5/internal/output"
 	"github.com/spf13/cobra"
 )
 
@@ -14,12 +14,11 @@ func newDeleteConsumerGroupCmd() *cobra.Command {
 		Aliases: []string{"consumer-groups"},
 		Short:   "delete a consumer-group",
 		Args:    cobra.MinimumNArgs(1),
-		Run: func(cmd *cobra.Command, args []string) {
-			if !k8s.NewOperation().TryRun(cmd, args) {
-				if err := (&consumergroups.ConsumerGroupOperation{}).DeleteConsumerGroups(args); err != nil {
-					output.Fail(err)
-				}
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if internal.IsKubernetesEnabled() {
+				return k8s.NewOperation().Run(cmd, args)
 			}
+			return (&consumergroups.ConsumerGroupOperation{}).DeleteConsumerGroups(args)
 		},
 		ValidArgsFunction: consumergroups.CompleteConsumerGroups,
 	}
