@@ -2,8 +2,8 @@ package deletion
 
 import (
 	"github.com/deviceinsight/kafkactl/v5/cmd/validation"
+	"github.com/deviceinsight/kafkactl/v5/internal"
 	"github.com/deviceinsight/kafkactl/v5/internal/k8s"
-	"github.com/deviceinsight/kafkactl/v5/internal/output"
 	"github.com/deviceinsight/kafkactl/v5/internal/topic"
 	"github.com/spf13/cobra"
 )
@@ -16,12 +16,11 @@ func newDeleteRecordsCmd() *cobra.Command {
 		Use:   "records TOPIC",
 		Short: "delete a records from a topic",
 		Args:  cobra.ExactArgs(1),
-		Run: func(cmd *cobra.Command, args []string) {
-			if !k8s.NewOperation().TryRun(cmd, args) {
-				if err := (&topic.Operation{}).DeleteRecords(args[0], flags); err != nil {
-					output.Fail(err)
-				}
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if internal.IsKubernetesEnabled() {
+				return k8s.NewOperation().Run(cmd, args)
 			}
+			return (&topic.Operation{}).DeleteRecords(args[0], flags)
 		},
 		ValidArgsFunction: topic.CompleteTopicNames,
 	}
