@@ -8,7 +8,7 @@ func NewDeleteCmd() *cobra.Command {
 
 	var cmdDelete = &cobra.Command{
 		Use:   "delete",
-		Short: "delete topics, consumerGroups, consumer-group-offset, acls, records",
+		Short: "delete topics, consumerGroups, consumer-group-offset, acls, records, users",
 	}
 
 	cmdDelete.AddCommand(newDeleteTopicCmd())
@@ -16,5 +16,6 @@ func NewDeleteCmd() *cobra.Command {
 	cmdDelete.AddCommand(newDeleteConsumerGroupOffsetCmd())
 	cmdDelete.AddCommand(newDeleteACLCmd())
 	cmdDelete.AddCommand(newDeleteRecordsCmd())
+	cmdDelete.AddCommand(newDeleteUserCmd())
 	return cmdDelete
 }
