@@ -1,11 +1,10 @@
 package describe
 
 import (
-	"strconv"
+	"github.com/deviceinsight/kafkactl/v5/internal"
 
 	"github.com/deviceinsight/kafkactl/v5/internal/broker"
 	"github.com/deviceinsight/kafkactl/v5/internal/k8s"
-	"github.com/deviceinsight/kafkactl/v5/internal/output"
 	"github.com/spf13/cobra"
 )
 
@@ -17,22 +16,16 @@ func newDescribeBrokerCmd() *cobra.Command {
 		Use:   "broker ID",
 		Short: "describe a broker",
 		Args:  cobra.MinimumNArgs(1),
-		Run: func(cmd *cobra.Command, args []string) {
-			if !k8s.NewOperation().TryRun(cmd, args) {
-
-				id, err := strconv.ParseInt(args[0], 10, 32)
-				if err != nil {
-					output.Fail(err)
-				}
-
-				if err := (&broker.Operation{}).DescribeBroker(int32(id), flags); err != nil {
-					output.Fail(err)
-				}
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if internal.IsKubernetesEnabled() {
+				return k8s.NewOperation().Run(cmd, args)
 			}
+			return (&broker.Operation{}).DescribeBroker(args[0], flags)
 		},
 		ValidArgsFunction: broker.CompleteBrokerIDs,
 	}
 
+	cmdDescribeBroker.Flags().BoolVarP(&flags.AllConfigs, "all-configs", "a", false, "print all configs including defaults")
 	cmdDescribeBroker.Flags().StringVarP(&flags.OutputFormat, "output", "o", flags.OutputFormat, "output format. One of: json|yaml|wide")
 
 	return cmdDescribeBroker
