@@ -2,7 +2,9 @@ package produce_test
 
 import (
 	"bufio"
+	"bytes"
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -10,15 +12,18 @@ import (
 	"strings"
 	"testing"
 
+	"google.golang.org/protobuf/encoding/protojson"
+	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/dynamicpb"
+
+	"github.com/riferrei/srclient"
+
+	"github.com/deviceinsight/kafkactl/v5/internal"
 	"github.com/deviceinsight/kafkactl/v5/internal/helpers/protobuf"
-
-	"github.com/jhump/protoreflect/dynamic"
-
 	"github.com/deviceinsight/kafkactl/v5/internal/testutil"
 )
 
 func TestProduceWithKeyAndValueIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "produce-topic")
@@ -39,7 +44,6 @@ func TestProduceWithKeyAndValueIntegration(t *testing.T) {
 }
 
 func TestProduceMessageWithHeadersIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "produce-topic")
@@ -60,7 +64,6 @@ func TestProduceMessageWithHeadersIntegration(t *testing.T) {
 }
 
 func TestProduceAvroMessageWithHeadersIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	valueSchema := `{
@@ -75,7 +78,7 @@ func TestProduceAvroMessageWithHeadersIntegration(t *testing.T) {
 }`
 	value := `{"name":"Peter Mueller"}`
 
-	topicName := testutil.CreateAvroTopic(t, "produce-topic", "", valueSchema)
+	topicName := testutil.CreateTopicWithSchema(t, "produce-topic", "", valueSchema, srclient.Avro)
 
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
 
@@ -92,8 +95,7 @@ func TestProduceAvroMessageWithHeadersIntegration(t *testing.T) {
 	testutil.AssertEquals(t, fmt.Sprintf("key1:value1,key\\:2:value\\:2#test-key#%s", value), kafkaCtl.GetStdOut())
 }
 
-func TestProduceAvroMessageOmitDefaultValue(t *testing.T) {
-
+func TestProduceAvroMessageOmitDefaultValueIntegration(t *testing.T) {
 	testutil.StartIntegrationTest(t)
 
 	valueSchema := `{
@@ -109,7 +111,7 @@ func TestProduceAvroMessageOmitDefaultValue(t *testing.T) {
 	 "CurrencyCode": "EUR"
 	}`
 
-	topicName := testutil.CreateAvroTopic(t, "produce-avro-topic", "", valueSchema)
+	topicName := testutil.CreateTopicWithSchema(t, "produce-avro-topic", "", valueSchema, srclient.Avro)
 
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
 
@@ -128,8 +130,7 @@ func TestProduceAvroMessageOmitDefaultValue(t *testing.T) {
 	testutil.AssertContainSubstring(t, `"ExpiresOn":null`, stdout)
 }
 
-func TestProduceAvroMessageWithUnionStandardJson(t *testing.T) {
-
+func TestProduceAvroMessageWithUnionStandardJsonIntegration(t *testing.T) {
 	testutil.StartIntegrationTest(t)
 
 	valueSchema := `{
@@ -147,7 +148,7 @@ func TestProduceAvroMessageWithUnionStandardJson(t *testing.T) {
 	 "ExpiresOn": "2022-12-12"
 	}`
 
-	topicName := testutil.CreateAvroTopic(t, "produce-topic", "", valueSchema)
+	topicName := testutil.CreateTopicWithSchema(t, "produce-topic", "", valueSchema, srclient.Avro)
 
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
 
@@ -166,8 +167,112 @@ func TestProduceAvroMessageWithUnionStandardJson(t *testing.T) {
 	testutil.AssertContainSubstring(t, `"ExpiresOn":"2022-12-12"`, stdout)
 }
 
-func TestProduceAvroMessageWithUnionAvroJson(t *testing.T) {
+func TestProduceRegistryProtobufMessageWithHeadersIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
 
+	valueSchema := `syntax = "proto3";
+  package foo.bar;
+
+  message Msg {
+    string name = 1;
+  }`
+	value := `{"name":"Peter Mueller"}`
+
+	topicName := testutil.CreateTopicWithSchema(t, "produce-protobuf-topic", "", valueSchema, srclient.Protobuf)
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "test-key", "--value", value, "-H", "key1:value1", "-H", "key\\:2:value\\:2"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit", "--print-keys", "--print-headers"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, fmt.Sprintf("key1:value1,key\\:2:value\\:2#test-key#%s", value), kafkaCtl.GetStdOut())
+}
+
+func TestProduceRegistryProtobufMessageOmitDefaultValueIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	valueSchema := `syntax = "proto3";
+  package foo.bar;
+
+  message Msg {
+    string current_code = 1;
+    string expires_on = 2;
+  }`
+	value := `{"currentCode":"EUR"}`
+
+	topicName := testutil.CreateTopicWithSchema(t, "produce-protobuf-topic", "", valueSchema, srclient.Protobuf)
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--value", value); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning",
+		"--proto-marshal-option", "emitDefaultValues", "--exit"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	stdout := kafkaCtl.GetStdOut()
+	testutil.AssertContainSubstring(t, `"currentCode":"EUR"`, stdout)
+	testutil.AssertContainSubstring(t, `"expiresOn":""`, stdout)
+}
+
+func TestProduceJsonMessageWithSchemaIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+
+	valueSchema := `{
+	  "$schema": "http://json-schema.org/draft-04/schema#",
+	  "type": "object",
+	  "properties": {
+		"CurrencyCode": {
+		  "type": "string"
+		},
+		"ExpiresOn": {
+		  "type": "string",
+		  "format": "date"
+		}
+	  },
+	  "required": [
+		"CurrencyCode",
+		"ExpiresOn"
+	  ]
+	}`
+
+	value := `{
+	 "CurrencyCode": "EUR",
+	 "ExpiresOn": "2022-12-12"
+	}`
+
+	topicName := testutil.CreateTopicWithSchema(t, "produce-topic", "", valueSchema, srclient.Json)
+
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--value", value); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	stdout := kafkaCtl.GetStdOut()
+	testutil.AssertContainSubstring(t, `"CurrencyCode": "EUR"`, stdout)
+	testutil.AssertContainSubstring(t, `"ExpiresOn": "2022-12-12"`, stdout)
+}
+
+func TestProduceAvroMessageWithUnionAvroJsonIntegration(t *testing.T) {
 	testutil.StartIntegrationTest(t)
 
 	valueSchema := `{
@@ -189,7 +294,7 @@ func TestProduceAvroMessageWithUnionAvroJson(t *testing.T) {
 		t.Fatalf("unable to set env variable: %v", err)
 	}
 
-	topicName := testutil.CreateAvroTopic(t, "produce-topic", "", valueSchema)
+	topicName := testutil.CreateTopicWithSchema(t, "produce-topic", "", valueSchema, srclient.Avro)
 
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
 
@@ -209,7 +314,6 @@ func TestProduceAvroMessageWithUnionAvroJson(t *testing.T) {
 }
 
 func TestProduceTombstoneIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "produce-topic")
@@ -231,7 +335,6 @@ func TestProduceTombstoneIntegration(t *testing.T) {
 }
 
 func TestProduceFromBase64Integration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "produce-topic")
@@ -254,7 +357,6 @@ func TestProduceFromBase64Integration(t *testing.T) {
 }
 
 func TestProduceFromHexIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	topicName := testutil.CreateTopic(t, "produce-topic")
@@ -277,7 +379,6 @@ func TestProduceFromHexIntegration(t *testing.T) {
 }
 
 func TestProduceAutoCompletionIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 
 	prefix := "produce-complete-"
@@ -337,28 +438,28 @@ func TestProduceProtoFileIntegration(t *testing.T) {
 		t.Fatalf("Failed to decode value: %s", err)
 	}
 
-	keyMessage := dynamic.NewMessage(protobuf.ResolveMessageType(protobuf.SearchContext{
+	keyMessage := dynamicpb.NewMessage(protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtoImportPaths: []string{protoPath},
 		ProtoFiles:       []string{"msg.proto"},
 	}, "TopicKey"))
-	valueMessage := dynamic.NewMessage(protobuf.ResolveMessageType(protobuf.SearchContext{
+	valueMessage := dynamicpb.NewMessage(protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtoImportPaths: []string{protoPath},
 		ProtoFiles:       []string{"msg.proto"},
 	}, "TopicMessage"))
 
-	if err = keyMessage.Unmarshal(rawKey); err != nil {
+	if err = proto.Unmarshal(rawKey, keyMessage); err != nil {
 		t.Fatalf("Unmarshal key failed: %s", err)
 	}
-	if err = valueMessage.Unmarshal(rawValue); err != nil {
+	if err = proto.Unmarshal(rawValue, valueMessage); err != nil {
 		t.Fatalf("Unmarshal value failed: %s", err)
 	}
 
-	actualKey, err := keyMessage.MarshalJSON()
+	actualKey, err := marshalJSON(keyMessage)
 	if err != nil {
 		t.Fatalf("Key to json failed: %s", err)
 	}
 
-	actualValue, err := valueMessage.MarshalJSON()
+	actualValue, err := marshalJSON(valueMessage)
 	if err != nil {
 		t.Fatalf("Value to json failed: %s", err)
 	}
@@ -368,7 +469,6 @@ func TestProduceProtoFileIntegration(t *testing.T) {
 }
 
 func TestProduceWithCSVFileIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 	topic := testutil.CreateTopic(t, "produce-topic-csv")
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
@@ -390,7 +490,6 @@ func TestProduceWithCSVFileIntegration(t *testing.T) {
 }
 
 func TestProduceWithCSVFileWithTimestampsFirstColumnIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 	topic := testutil.CreateTopic(t, "produce-topic-csv")
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
@@ -412,7 +511,6 @@ func TestProduceWithCSVFileWithTimestampsFirstColumnIntegration(t *testing.T) {
 }
 
 func TestProduceWithCSVFileWithTimestampsSecondColumnIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 	topic := testutil.CreateTopic(t, "produce-topic-csv")
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
@@ -434,7 +532,6 @@ func TestProduceWithCSVFileWithTimestampsSecondColumnIntegration(t *testing.T) {
 }
 
 func TestProduceWithJSONFileIntegration(t *testing.T) {
-
 	testutil.StartIntegrationTest(t)
 	topic := testutil.CreateTopic(t, "produce-topic-json")
 	kafkaCtl := testutil.CreateKafkaCtlCommand()
@@ -447,13 +544,37 @@ func TestProduceWithJSONFileIntegration(t *testing.T) {
 		t.Fatalf("failed to execute command: %v", err)
 	}
 
-	testutil.AssertEquals(t, "3 messages produced", kafkaCtl.GetStdOut())
+	testutil.AssertEquals(t, "6 messages produced", kafkaCtl.GetStdOut())
 
-	if _, err := kafkaCtl.Execute("consume", topic, "--from-beginning", "--print-keys", "--exit"); err != nil {
+	if _, err := kafkaCtl.Execute("consume", topic, "--from-beginning", "--print-keys", "--print-headers", "--exit"); err != nil {
 		t.Fatalf("failed to execute command: %v", err)
 	}
 
-	testutil.AssertEquals(t, "1#a\n2#b\n3#c", kafkaCtl.GetStdOut())
+	expectedMessages := []string{"a:b,c:1#1#a", "#2#b", "x:y#3#c", "##value-only", "#key-only#null", "##null"}
+	testutil.AssertArraysEquals(t, expectedMessages, kafkaCtl.GetStdOutLines())
+}
+
+func TestProduceWithJSONFileBase64ValuesIntegration(t *testing.T) {
+	testutil.StartIntegrationTest(t)
+	topic := testutil.CreateTopic(t, "produce-topic-json-base64-values")
+	kafkaCtl := testutil.CreateKafkaCtlCommand()
+
+	dataFilePath := filepath.Join(testutil.RootDir, "internal", "testutil", "testdata")
+
+	if _, err := kafkaCtl.Execute("produce", topic,
+		"--file", filepath.Join(dataFilePath, "msg-base64.json"),
+		"--value-encoding", "base64",
+		"--input-format", "json"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, "3 messages produced", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topic, "--from-beginning", "--print-keys", "--value-encoding", "hex", "--exit"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	testutil.AssertEquals(t, "1#000000000001\n2#68656c6c6f\n3#6b61666b61", kafkaCtl.GetStdOut())
 }
 
 func TestProduceProtoFileWithOnlyKeyEncodedIntegration(t *testing.T) {
@@ -487,16 +608,16 @@ func TestProduceProtoFileWithOnlyKeyEncodedIntegration(t *testing.T) {
 		t.Fatalf("Failed to decode key: %s", err)
 	}
 
-	keyMessage := dynamic.NewMessage(protobuf.ResolveMessageType(protobuf.SearchContext{
+	keyMessage := dynamicpb.NewMessage(protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtoImportPaths: []string{protoPath},
 		ProtoFiles:       []string{"msg.proto"},
 	}, "TopicKey"))
 
-	if err = keyMessage.Unmarshal(rawKey); err != nil {
+	if err = proto.Unmarshal(rawKey, keyMessage); err != nil {
 		t.Fatalf("Unmarshal key failed: %s", err)
 	}
 
-	actualKey, err := keyMessage.MarshalJSON()
+	actualKey, err := marshalJSON(keyMessage)
 	if err != nil {
 		t.Fatalf("Key to json failed: %s", err)
 	}
@@ -544,28 +665,28 @@ func TestProduceProtoFileWithoutProtoImportPathIntegration(t *testing.T) {
 		t.Fatalf("Failed to decode value: %s", err)
 	}
 
-	keyMessage := dynamic.NewMessage(protobuf.ResolveMessageType(protobuf.SearchContext{
+	keyMessage := dynamicpb.NewMessage(protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtoImportPaths: []string{protoPath},
 		ProtoFiles:       []string{"msg.proto"},
 	}, "TopicKey"))
-	valueMessage := dynamic.NewMessage(protobuf.ResolveMessageType(protobuf.SearchContext{
+	valueMessage := dynamicpb.NewMessage(protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtoImportPaths: []string{protoPath},
 		ProtoFiles:       []string{"msg.proto"},
 	}, "TopicMessage"))
 
-	if err = keyMessage.Unmarshal(rawKey); err != nil {
+	if err = proto.Unmarshal(rawKey, keyMessage); err != nil {
 		t.Fatalf("Unmarshal key failed: %s", err)
 	}
-	if err = valueMessage.Unmarshal(rawValue); err != nil {
+	if err = proto.Unmarshal(rawValue, valueMessage); err != nil {
 		t.Fatalf("Unmarshal value failed: %s", err)
 	}
 
-	actualKey, err := keyMessage.MarshalJSON()
+	actualKey, err := marshalJSON(keyMessage)
 	if err != nil {
 		t.Fatalf("Key to json failed: %s", err)
 	}
 
-	actualValue, err := valueMessage.MarshalJSON()
+	actualValue, err := marshalJSON(valueMessage)
 	if err != nil {
 		t.Fatalf("Value to json failed: %s", err)
 	}
@@ -611,26 +732,26 @@ func TestProduceProtosetFileIntegration(t *testing.T) {
 		t.Fatalf("Failed to decode value: %s", err)
 	}
 
-	keyMessage := dynamic.NewMessage(protobuf.ResolveMessageType(protobuf.SearchContext{
+	keyMessage := dynamicpb.NewMessage(protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtosetFiles: []string{protoPath},
 	}, "TopicKey"))
-	valueMessage := dynamic.NewMessage(protobuf.ResolveMessageType(protobuf.SearchContext{
+	valueMessage := dynamicpb.NewMessage(protobuf.ResolveMessageType(internal.ProtobufConfig{
 		ProtosetFiles: []string{protoPath},
 	}, "TopicMessage"))
 
-	if err = keyMessage.Unmarshal(rawKey); err != nil {
+	if err = proto.Unmarshal(rawKey, keyMessage); err != nil {
 		t.Fatalf("Unmarshal key failed: %s", err)
 	}
-	if err = valueMessage.Unmarshal(rawValue); err != nil {
+	if err = proto.Unmarshal(rawValue, valueMessage); err != nil {
 		t.Fatalf("Unmarshal value failed: %s", err)
 	}
 
-	actualKey, err := keyMessage.MarshalJSON()
+	actualKey, err := marshalJSON(keyMessage)
 	if err != nil {
 		t.Fatalf("Key to json failed: %s", err)
 	}
 
-	actualValue, err := valueMessage.MarshalJSON()
+	actualValue, err := marshalJSON(valueMessage)
 	if err != nil {
 		t.Fatalf("Value to json failed: %s", err)
 	}
@@ -737,4 +858,20 @@ func TestProduceLongMessageFailsIntegration(t *testing.T) {
 	} else {
 		t.Fatalf("Expected producer to fail")
 	}
+}
+
+func marshalJSON(message *dynamicpb.Message) ([]byte, error) {
+	jsonValue, err := protojson.MarshalOptions{Indent: ""}.Marshal(message)
+	if err != nil {
+		return nil, err
+	}
+
+	// this is needed to eliminate whitespace randomization
+	// https://github.com/golang/protobuf/issues/1082
+	buffer := new(bytes.Buffer)
+	if err := json.Compact(buffer, jsonValue); err != nil {
+		return nil, err
+	}
+
+	return buffer.Bytes(), nil
 }
