@@ -2,8 +2,10 @@ package cmd
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/deviceinsight/kafkactl/v5/internal/global"
+	"github.com/hashicorp/go-plugin"
 
 	"github.com/deviceinsight/kafkactl/v5/cmd/alter"
 	"github.com/deviceinsight/kafkactl/v5/cmd/attach"
@@ -24,14 +26,22 @@ import (
 func NewKafkactlCommand(streams output.IOStreams) *cobra.Command {
 
 	var rootCmd = &cobra.Command{
-		Use:   "kafkactl",
-		Short: "command-line interface for Apache Kafka",
-		Long:  `A command-line interface the simplifies interaction with Kafka.`,
+		Use:           "kafkactl",
+		Short:         "command-line interface for Apache Kafka",
+		Long:          `A command-line interface the simplifies interaction with Kafka.`,
+		SilenceUsage:  true,
+		SilenceErrors: true,
 	}
 
 	globalConfig := global.NewConfig()
 
-	cobra.OnInitialize(globalConfig.Init)
+	cobra.OnInitialize(func() {
+		if err := globalConfig.Init(); err != nil {
+			output.Warnf("failed to initialize config: %v", err)
+			os.Exit(1)
+		}
+	})
+	cobra.OnFinalize(plugin.CleanupClients)
 
 	rootCmd.AddCommand(config.NewConfigCmd())
 	rootCmd.AddCommand(consume.NewConsumeCmd())
@@ -54,6 +64,14 @@ func NewKafkactlCommand(streams output.IOStreams) *cobra.Command {
 	rootCmd.PersistentFlags().StringVarP(&globalFlags.ConfigFile, "config-file", "C", "",
 		fmt.Sprintf("config file. default locations: %v", globalConfig.DefaultPaths()))
 	rootCmd.PersistentFlags().BoolVarP(&globalFlags.Verbose, "verbose", "V", false, "verbose output")
+	rootCmd.PersistentFlags().StringVar(&globalFlags.Context, "context", "", "The name of the context to use")
+
+	err := rootCmd.RegisterFlagCompletionFunc("context", func(_ *cobra.Command, _ []string, _ string) ([]string, cobra.ShellCompDirective) {
+		return global.ListAvailableContexts(), cobra.ShellCompDirectiveNoFileComp
+	})
+	if err != nil {
+		panic(err)
+	}
 
 	k8s.KafkaCtlVersion = Version
 
