@@ -10,23 +10,20 @@ import (
 	"strings"
 	"time"
 
+	"github.com/deviceinsight/kafkactl/v5/internal/helpers/avro"
+
 	"github.com/deviceinsight/kafkactl/v5/internal/auth"
 
 	"github.com/deviceinsight/kafkactl/v5/internal/global"
 
-	"github.com/deviceinsight/kafkactl/v5/internal/helpers/avro"
-
 	"github.com/IBM/sarama"
 	"github.com/deviceinsight/kafkactl/v5/internal/helpers"
-	"github.com/deviceinsight/kafkactl/v5/internal/helpers/protobuf"
 	"github.com/deviceinsight/kafkactl/v5/internal/output"
 	"github.com/pkg/errors"
 	"github.com/spf13/viper"
 )
 
-var (
-	invalidClientIDCharactersRegExp = regexp.MustCompile(`[^a-zA-Z0-9_-]`)
-)
+var invalidClientIDCharactersRegExp = regexp.MustCompile(`[^a-zA-Z0-9-]+`)
 
 type TokenProvider struct {
 	PluginName string
@@ -39,6 +36,34 @@ type SaslConfig struct {
 	Password      string
 	Mechanism     string
 	TokenProvider TokenProvider
+	Version       string
+}
+
+type SchemaRegistryConfig struct {
+	URL            string
+	RequestTimeout time.Duration
+	TLS            TLSConfig
+	Username       string
+	Password       string
+}
+
+type ProtobufConfig struct {
+	ProtosetFiles    []string
+	ProtoFiles       []string
+	ProtoImportPaths []string
+	MarshalOptions   ProtobufMarshalOptions
+}
+
+type ProtobufMarshalOptions struct {
+	AllowPartial      bool
+	UseProtoNames     bool
+	UseEnumNumbers    bool
+	EmitUnpopulated   bool
+	EmitDefaultValues bool
+}
+
+type AvroConfig struct {
+	JSONCodec avro.JSONCodec
 }
 
 type TLSConfig struct {
@@ -49,6 +74,13 @@ type TLSConfig struct {
 	Insecure bool
 }
 
+type K8sToleration struct {
+	Key      string `json:"key" yaml:"key"`
+	Operator string `json:"operator" yaml:"operator"`
+	Value    string `json:"value" yaml:"value"`
+	Effect   string `json:"effect" yaml:"effect"`
+}
+
 type K8sConfig struct {
 	Enabled         bool
 	Binary          string
@@ -57,11 +89,16 @@ type K8sConfig struct {
 	Namespace       string
 	Image           string
 	ImagePullSecret string
+	TLSSecret       string
 	ServiceAccount  string
+	AsUser          string
 	KeepPod         bool
 	Labels          map[string]string
 	Annotations     map[string]string
 	NodeSelector    map[string]string
+	Affinity        map[string]any
+	Resources       map[string]any
+	Tolerations     []K8sToleration
 }
 
 type ConsumerConfig struct {
@@ -72,22 +109,24 @@ type ProducerConfig struct {
 	Partitioner     string
 	RequiredAcks    string
 	MaxMessageBytes int
+	ValueSerializer string
+	KeySerializer   string
 }
 
 type ClientContext struct {
-	Name               string
-	Brokers            []string
-	TLS                TLSConfig
-	Sasl               SaslConfig
-	Kubernetes         K8sConfig
-	RequestTimeout     time.Duration
-	ClientID           string
-	KafkaVersion       sarama.KafkaVersion
-	AvroSchemaRegistry string
-	AvroJSONCodec      avro.JSONCodec
-	Protobuf           protobuf.SearchContext
-	Producer           ProducerConfig
-	Consumer           ConsumerConfig
+	Name           string
+	Brokers        []string
+	TLS            TLSConfig
+	Sasl           SaslConfig
+	Kubernetes     K8sConfig
+	RequestTimeout time.Duration
+	ClientID       string
+	KafkaVersion   sarama.KafkaVersion
+	Avro           AvroConfig
+	Protobuf       ProtobufConfig
+	SchemaRegistry SchemaRegistryConfig
+	Producer       ProducerConfig
+	Consumer       ConsumerConfig
 }
 
 type Config struct {
@@ -97,8 +136,12 @@ type Config struct {
 
 func CreateClientContext() (ClientContext, error) {
 	var context ClientContext
+	var err error
 
-	context.Name = global.GetCurrentContext()
+	context.Name, err = global.GetCurrentContext()
+	if err != nil {
+		return context, err
+	}
 
 	if viper.Get("contexts."+context.Name) == nil {
 		return context, errors.Errorf("no context with name %s found", context.Name)
@@ -113,9 +156,15 @@ func CreateClientContext() (ClientContext, error) {
 
 	context.Brokers = viper.GetStringSlice("contexts." + context.Name + ".brokers")
 	context.TLS.Enabled = viper.GetBool("contexts." + context.Name + ".tls.enabled")
-	context.TLS.CA = viper.GetString("contexts." + context.Name + ".tls.ca")
-	context.TLS.Cert = viper.GetString("contexts." + context.Name + ".tls.cert")
-	context.TLS.CertKey = viper.GetString("contexts." + context.Name + ".tls.certKey")
+	if context.TLS.CA, err = resolvePath("contexts." + context.Name + ".tls.ca"); err != nil {
+		return context, err
+	}
+	if context.TLS.Cert, err = resolvePath("contexts." + context.Name + ".tls.cert"); err != nil {
+		return context, err
+	}
+	if context.TLS.CertKey, err = resolvePath("contexts." + context.Name + ".tls.certKey"); err != nil {
+		return context, err
+	}
 	context.TLS.Insecure = viper.GetBool("contexts." + context.Name + ".tls.insecure")
 	context.ClientID = viper.GetString("contexts." + context.Name + ".clientID")
 
@@ -126,37 +175,78 @@ func CreateClientContext() (ClientContext, error) {
 	} else {
 		return context, err
 	}
-	context.AvroSchemaRegistry = viper.GetString("contexts." + context.Name + ".avro.schemaRegistry")
-	context.AvroJSONCodec = avro.ParseJSONCodec(viper.GetString("contexts." + context.Name + ".avro.jsonCodec"))
-	context.Protobuf.ProtosetFiles = viper.GetStringSlice("contexts." + context.Name + ".protobuf.protosetFiles")
-	context.Protobuf.ProtoImportPaths = viper.GetStringSlice("contexts." + context.Name + ".protobuf.importPaths")
+	context.Avro.JSONCodec = avro.ParseJSONCodec(viper.GetString("contexts." + context.Name + ".avro.jsonCodec"))
+	context.SchemaRegistry.URL = viper.GetString("contexts." + context.Name + ".schemaRegistry.url")
+	context.SchemaRegistry.RequestTimeout = viper.GetDuration("contexts." + context.Name + ".schemaRegistry.requestTimeout")
+	context.SchemaRegistry.TLS.Enabled = viper.GetBool("contexts." + context.Name + ".schemaRegistry.tls.enabled")
+	if context.SchemaRegistry.TLS.CA, err = resolvePath("contexts." + context.Name + ".schemaRegistry.tls.ca"); err != nil {
+		return context, err
+	}
+	if context.SchemaRegistry.TLS.Cert, err = resolvePath("contexts." + context.Name + ".schemaRegistry.tls.cert"); err != nil {
+		return context, err
+	}
+	if context.SchemaRegistry.TLS.CertKey, err = resolvePath("contexts." + context.Name + ".schemaRegistry.tls.certKey"); err != nil {
+		return context, err
+	}
+	context.SchemaRegistry.TLS.Insecure = viper.GetBool("contexts." + context.Name + ".schemaRegistry.tls.insecure")
+	context.SchemaRegistry.Username = viper.GetString("contexts." + context.Name + ".schemaRegistry.username")
+	context.SchemaRegistry.Password = viper.GetString("contexts." + context.Name + ".schemaRegistry.password")
+	if context.Protobuf.ProtosetFiles, err = resolvePaths("contexts." + context.Name + ".protobuf.protosetFiles"); err != nil {
+		return context, err
+	}
+	if context.Protobuf.ProtoImportPaths, err = resolvePaths("contexts." + context.Name + ".protobuf.importPaths"); err != nil {
+		return context, err
+	}
 	context.Protobuf.ProtoFiles = viper.GetStringSlice("contexts." + context.Name + ".protobuf.protoFiles")
 	context.Producer.Partitioner = viper.GetString("contexts." + context.Name + ".producer.partitioner")
 	context.Producer.RequiredAcks = viper.GetString("contexts." + context.Name + ".producer.requiredAcks")
 	context.Producer.MaxMessageBytes = viper.GetInt("contexts." + context.Name + ".producer.maxMessageBytes")
+	context.Producer.ValueSerializer = viper.GetString("contexts." + context.Name + ".producer.valueSerializer")
+	context.Producer.KeySerializer = viper.GetString("contexts." + context.Name + ".producer.keySerializer")
 	context.Consumer.IsolationLevel = viper.GetString("contexts." + context.Name + ".consumer.isolationLevel")
 	context.Sasl.Enabled = viper.GetBool("contexts." + context.Name + ".sasl.enabled")
 	context.Sasl.Username = viper.GetString("contexts." + context.Name + ".sasl.username")
 	context.Sasl.Password = viper.GetString("contexts." + context.Name + ".sasl.password")
 	context.Sasl.Mechanism = viper.GetString("contexts." + context.Name + ".sasl.mechanism")
+	context.Sasl.Version = viper.GetString("contexts." + context.Name + ".sasl.version")
 	context.Sasl.TokenProvider.PluginName = viper.GetString("contexts." + context.Name + ".sasl.tokenProvider.plugin")
 	context.Sasl.TokenProvider.Options = viper.GetStringMap("contexts." + context.Name + ".sasl.tokenProvider.options")
 
 	viper.SetDefault("contexts."+context.Name+".kubernetes.binary", "kubectl")
-	context.Kubernetes.Enabled = viper.GetBool("contexts." + context.Name + ".kubernetes.enabled")
-	context.Kubernetes.Binary = viper.GetString("contexts." + context.Name + ".kubernetes.binary")
-	context.Kubernetes.KubeConfig = viper.GetString("contexts." + context.Name + ".kubernetes.kubeConfig")
+	context.Kubernetes.Enabled = IsKubernetesEnabled()
+
+	if context.Kubernetes.Enabled {
+		if context.Kubernetes.Binary, err = resolvePath("contexts." + context.Name + ".kubernetes.binary"); err != nil {
+			return context, err
+		}
+		if context.Kubernetes.KubeConfig, err = resolvePath("contexts." + context.Name + ".kubernetes.kubeConfig"); err != nil {
+			return context, err
+		}
+	}
 	context.Kubernetes.KubeContext = viper.GetString("contexts." + context.Name + ".kubernetes.kubeContext")
 	context.Kubernetes.Namespace = viper.GetString("contexts." + context.Name + ".kubernetes.namespace")
 	context.Kubernetes.Image = viper.GetString("contexts." + context.Name + ".kubernetes.image")
 	context.Kubernetes.ImagePullSecret = viper.GetString("contexts." + context.Name + ".kubernetes.imagePullSecret")
+	context.Kubernetes.TLSSecret = viper.GetString("contexts." + context.Name + ".kubernetes.tlsSecret")
 	context.Kubernetes.ServiceAccount = viper.GetString("contexts." + context.Name + ".kubernetes.serviceAccount")
+	context.Kubernetes.AsUser = viper.GetString("contexts." + context.Name + ".kubernetes.asUser")
 	context.Kubernetes.KeepPod = viper.GetBool("contexts." + context.Name + ".kubernetes.keepPod")
 	context.Kubernetes.Labels = viper.GetStringMapString("contexts." + context.Name + ".kubernetes.labels")
 	context.Kubernetes.Annotations = viper.GetStringMapString("contexts." + context.Name + ".kubernetes.annotations")
 	context.Kubernetes.NodeSelector = viper.GetStringMapString("contexts." + context.Name + ".kubernetes.nodeSelector")
+	context.Kubernetes.Affinity = viper.GetStringMap("contexts." + context.Name + ".kubernetes.affinity")
+	context.Kubernetes.Resources = viper.GetStringMap("contexts." + context.Name + ".kubernetes.resources")
+
+	if err := viper.UnmarshalKey("contexts."+context.Name+".kubernetes.tolerations", &context.Kubernetes.Tolerations); err != nil {
+		return context, err
+	}
 
 	return context, nil
+}
+
+func IsKubernetesEnabled() bool {
+	contextName, _ := global.GetCurrentContext()
+	return viper.GetBool("contexts." + contextName + ".kubernetes.enabled")
 }
 
 func CreateClient(context *ClientContext) (sarama.Client, error) {
@@ -176,8 +266,7 @@ func CreateClusterAdmin(context *ClientContext) (sarama.ClusterAdmin, error) {
 }
 
 func CreateClientConfig(context *ClientContext) (*sarama.Config, error) {
-
-	var config = sarama.NewConfig()
+	config := sarama.NewConfig()
 	config.Version = context.KafkaVersion
 	config.ClientID = GetClientID(context, "kafkactl-")
 
@@ -208,6 +297,12 @@ func CreateClientConfig(context *ClientContext) (*sarama.Config, error) {
 		config.Net.SASL.Enable = true
 		config.Net.SASL.User = context.Sasl.Username
 		config.Net.SASL.Password = context.Sasl.Password
+		if strings.EqualFold(context.Sasl.Version, "v0") {
+			config.Net.SASL.Version = sarama.SASLHandshakeV0
+		}
+		if strings.EqualFold(context.Sasl.Version, "v1") {
+			config.Net.SASL.Version = sarama.SASLHandshakeV1
+		}
 		switch context.Sasl.Mechanism {
 		case "scram-sha512":
 			config.Net.SASL.Mechanism = sarama.SASLTypeSCRAMSHA512
@@ -242,7 +337,6 @@ func CreateClientConfig(context *ClientContext) (*sarama.Config, error) {
 }
 
 func GetClientID(context *ClientContext, defaultPrefix string) string {
-
 	var (
 		err error
 		usr *user.User
@@ -257,19 +351,51 @@ func GetClientID(context *ClientContext, defaultPrefix string) string {
 	return defaultPrefix + sanitizeUsername(usr.Username)
 }
 
+func resolvePaths(key string) ([]string, error) {
+	filenames := viper.GetStringSlice(key)
+	resolved := make([]string, len(filenames))
+	var err error
+
+	for i, filename := range filenames {
+		resolved[i], err = global.ResolvePath(filename)
+		if err != nil {
+			return nil, fmt.Errorf("failed to resolve path for key %q: %w", key, err)
+		}
+	}
+
+	return resolved, nil
+}
+
+func resolvePath(key string) (string, error) {
+	filename := viper.GetString(key)
+
+	if filename == "" {
+		return filename, nil
+	}
+
+	if IsKubernetesEnabled() {
+		return filename, nil
+	}
+
+	resolved, err := global.ResolvePath(filename)
+	if err != nil {
+		return resolved, fmt.Errorf("failed to resolve path for key %q: %w", key, err)
+	}
+	return resolved, nil
+}
+
 func sanitizeUsername(u string) string {
 	// Windows user may have format "DOMAIN|MACHINE\username", remove domain/machine if present
 	s := strings.Split(u, "\\")
 	u = s[len(s)-1]
 	// Windows account can contain spaces or other special characters not supported
 	// in client ID. Keep the bare minimum and ditch the rest.
-	return invalidClientIDCharactersRegExp.ReplaceAllString(u, "")
+	return invalidClientIDCharactersRegExp.ReplaceAllString(u, "-")
 }
 
 // setupTlsConfig takes the paths to a tls certificate, CA, and certificate key in
 // a PEM format and returns a constructed tls.Config object.
 func setupTLSConfig(tlsConfig TLSConfig) (*tls.Config, error) {
-
 	if !tlsConfig.Enabled {
 		return nil, errors.Errorf("tls should be enabled at this point")
 	}
@@ -330,7 +456,6 @@ func kafkaVersion(s string) (sarama.KafkaVersion, error) {
 }
 
 func TopicExists(client *sarama.Client, name string) (bool, error) {
-
 	var (
 		err    error
 		topics []string
@@ -350,7 +475,6 @@ func TopicExists(client *sarama.Client, name string) (bool, error) {
 }
 
 func ListConfigs(admin *sarama.ClusterAdmin, resource sarama.ConfigResource, includeDefaults bool) ([]Config, error) {
-
 	var (
 		configEntries []sarama.ConfigEntry
 		err           error
@@ -364,10 +488,9 @@ func ListConfigs(admin *sarama.ClusterAdmin, resource sarama.ConfigResource, inc
 }
 
 func listConfigsFromEntries(configEntries []sarama.ConfigEntry, includeDefaults bool) []Config {
-	var configs = make([]Config, 0)
+	configs := make([]Config, 0)
 
 	for _, configEntry := range configEntries {
-
 		if includeDefaults || (!configEntry.Default && configEntry.Source != sarama.SourceDefault) {
 			entry := Config{Name: configEntry.Name, Value: configEntry.Value}
 			configs = append(configs, entry)
