@@ -2,9 +2,13 @@ package global
 
 import (
 	"errors"
+	"fmt"
 	"os"
+	"os/exec"
+	"path"
 	"path/filepath"
 	"runtime"
+	"sort"
 	"strings"
 
 	"github.com/deviceinsight/kafkactl/v5/internal/output"
@@ -13,11 +17,17 @@ import (
 
 type Flags struct {
 	ConfigFile string
+	Context    string
 	Verbose    bool
 }
 
 const defaultContextPrefix = "CONTEXTS_DEFAULT_"
-const GoContextKey = "global-config"
+
+const configFileName = "config"
+const writableConfigFileName = "current-context"
+
+const ConfigEnvVariable = "KAFKA_CTL_CONFIG"
+const WritableConfigEnvVariable = "KAFKA_CTL_WRITABLE_CONFIG"
 
 var projectConfigNames = []string{"kafkactl.yml", ".kafkactl.yml"}
 
@@ -25,8 +35,6 @@ var configPaths = []string{
 	"$HOME/.config/kafkactl",
 	"$HOME/.kafkactl",
 	"$APPDATA/kafkactl",
-	"$SNAP_REAL_HOME/.config/kafkactl",
-	"$SNAP_DATA/kafkactl",
 	"/etc/kafkactl",
 }
 
@@ -35,9 +43,10 @@ var configInstance *config
 type Config interface {
 	Flags() *Flags
 	DefaultPaths() []string
-	Init()
+	Init() error
 	currentContext() string
 	setCurrentContext(contextName string) error
+	SetWritableConfig(viper *viper.Viper)
 }
 
 func NewConfig() Config {
@@ -51,12 +60,84 @@ func GetFlags() Flags {
 	return configInstance.flags
 }
 
-func GetCurrentContext() string {
-	return configInstance.currentContext()
+func ListAvailableContexts() []string {
+
+	var contexts []string
+	for k := range viper.GetStringMap("contexts") {
+		contexts = append(contexts, k)
+	}
+
+	sort.Strings(contexts)
+	return contexts
+}
+
+func GetCurrentContext() (string, error) {
+	var context = configInstance.Flags().Context
+	if context != "" {
+		contexts := viper.GetStringMap("contexts")
+
+		// check if it is an existing context
+		if _, ok := contexts[context]; !ok {
+			return "", fmt.Errorf("not a valid context: %s", context)
+		}
+
+		return context, nil
+	}
+
+	return configInstance.currentContext(), nil
 }
 
 func SetCurrentContext(contextName string) error {
 	return configInstance.setCurrentContext(contextName)
+}
+
+func ResolvePath(filename string) (string, error) {
+
+	wd, _ := os.Getwd()
+	searchPaths := map[string]bool{wd: true}
+
+	if _, err := os.Stat(filename); err == nil {
+		return filename, err
+	} else if !errors.Is(err, os.ErrNotExist) || filepath.IsAbs(filename) {
+		return "", fmt.Errorf("unable to resolve path %q: %v", filename, err)
+	}
+
+	// relative to config file
+	configDir := path.Dir(viper.ConfigFileUsed())
+	if configDir != "." {
+		searchPaths[configDir] = true
+		cfgFilename := filepath.Join(path.Dir(viper.ConfigFileUsed()), filename)
+		if _, err := os.Stat(cfgFilename); err == nil {
+			return cfgFilename, err
+		} else if !errors.Is(err, os.ErrNotExist) {
+			return "", fmt.Errorf("unable to resolve path %q: %v", cfgFilename, err)
+		}
+	}
+
+	// relative to EXTRA_PATHS
+	if os.Getenv("EXTRA_PATHS") != "" {
+		for _, extraPath := range strings.Split(os.Getenv("EXTRA_PATHS"), ":") {
+			searchPaths[extraPath] = true
+			cfgFilename := filepath.Join(extraPath, filename)
+			if _, err := os.Stat(cfgFilename); err == nil {
+				return cfgFilename, err
+			} else if !errors.Is(err, os.ErrNotExist) {
+				return "", fmt.Errorf("unable to resolve path %q: %v", cfgFilename, err)
+			}
+		}
+	}
+
+	// check in path (e.g. for kubectl binary)
+	if _, err := exec.LookPath(filename); err == nil {
+		return filename, nil
+	}
+
+	locations := ""
+	for searchPath := range searchPaths {
+		locations += searchPath + ","
+	}
+
+	return "", fmt.Errorf("cannot find %q in locations: [%s]", filename, strings.Trim(locations, ","))
 }
 
 type config struct {
@@ -75,13 +156,18 @@ func (c *config) DefaultPaths() []string {
 func (c *config) currentContext() string {
 	return c.writableConfig.GetString("current-context")
 }
+
 func (c *config) setCurrentContext(contextName string) error {
 	c.writableConfig.Set("current-context", contextName)
 	return c.writableConfig.WriteConfig()
 }
 
+func (c *config) SetWritableConfig(viper *viper.Viper) {
+	c.writableConfig = viper
+}
+
 // Init reads in config file and ENV variables if set.
-func (c *config) Init() {
+func (c *config) Init() error {
 
 	viper.Reset()
 
@@ -94,52 +180,45 @@ func (c *config) Init() {
 	switch {
 	case c.flags.ConfigFile != "":
 		configFile = &c.flags.ConfigFile
-	case os.Getenv("KAFKA_CTL_CONFIG") != "":
-		envConfig := os.Getenv("KAFKA_CTL_CONFIG")
+	case os.Getenv(ConfigEnvVariable) != "":
+		envConfig := os.Getenv(ConfigEnvVariable)
 		configFile = &envConfig
-	}
-
-	if c.flags.Verbose && os.Getenv("SNAP_NAME") != "" {
-		output.Debugf("Running snap version %s on %s", os.Getenv("SNAP_VERSION"), os.Getenv("SNAP_ARCH"))
 	}
 
 	mapEnvVariables()
 
-	if err := c.loadConfig(viper.GetViper(), configFile); err != nil {
+	if err := loadConfig(configFileName, viper.GetViper(), configFile); err != nil {
 		if isUnknownError(err) {
-			output.Failf("Error reading config file: %s (%v)", viper.ConfigFileUsed(), err.Error())
+			return fmt.Errorf("error reading config file: %s (%v)", viper.ConfigFileUsed(), err.Error())
 		}
-		err = generateDefaultConfig()
+		err = generateDefaultConfig(configFileName, viper.GetViper())
 		if err != nil {
-			output.Failf("Error generating default config file: %v", err.Error())
-		}
-
-		// We read generated config now
-		if err = c.loadConfig(viper.GetViper(), configFile); err != nil {
-			output.Failf("Error reading config file: %s (%v)", viper.ConfigFileUsed(), err.Error())
+			return fmt.Errorf("error generating default config file: %v", err.Error())
 		}
 	}
 
-	if configFile != nil && viper.GetString("current-context") == "" {
-		// assuming the provided configFile is read-only
-		c.writableConfig = viper.New()
-		if err := c.loadConfig(c.writableConfig, nil); err != nil {
-			if isUnknownError(err) {
-				output.Failf("Error reading config file: %s (%v)", c.writableConfig.ConfigFileUsed(), err.Error())
-			}
-			err = generateDefaultConfig()
-			if err != nil {
-				output.Failf("Error generating default config file: %v", err.Error())
-			}
+	var writableConfigFile *string
 
-			// We read generated config now
-			if err = c.loadConfig(c.writableConfig, configFile); err != nil {
-				output.Failf("Error reading config file: %s (%v)", viper.ConfigFileUsed(), err.Error())
-			}
-		}
-	} else {
-		c.writableConfig = viper.GetViper()
+	if os.Getenv(WritableConfigEnvVariable) != "" {
+		envConfig := os.Getenv(WritableConfigEnvVariable)
+		writableConfigFile = &envConfig
+	} else if os.Getenv(ConfigEnvVariable) != "" {
+		// use config file provided via env as base path
+		envConfig := filepath.Join(path.Dir(os.Getenv(ConfigEnvVariable)), writableConfigFileName+".yml")
+		writableConfigFile = &envConfig
 	}
+
+	c.writableConfig = viper.New()
+	if err := loadConfig(writableConfigFileName, c.writableConfig, writableConfigFile); err != nil {
+		if isUnknownError(err) {
+			return fmt.Errorf("error reading config file: %s (%v)", c.writableConfig.ConfigFileUsed(), err.Error())
+		}
+		err = generateDefaultConfig(writableConfigFileName, c.writableConfig)
+		if err != nil {
+			return fmt.Errorf("error generating default config file: %v", err.Error())
+		}
+	}
+	return nil
 }
 
 func isUnknownError(err error) bool {
@@ -152,15 +231,15 @@ func isUnknownError(err error) bool {
 	return !isConfigFileNotFoundError && !isOsPathError
 }
 
-func (c *config) loadConfig(viperInstance *viper.Viper, configFile *string) error {
+func loadConfig(name string, viperInstance *viper.Viper, configFile *string) error {
 
 	if configFile != nil {
 		viperInstance.SetConfigFile(*configFile)
 	} else {
-		for _, path := range configPaths {
-			viperInstance.AddConfigPath(os.ExpandEnv(path))
+		for _, p := range configPaths {
+			viperInstance.AddConfigPath(os.ExpandEnv(p))
 		}
-		viperInstance.SetConfigName("config")
+		viperInstance.SetConfigName(name)
 	}
 
 	replacer := strings.NewReplacer("-", "_", ".", "_")
@@ -187,31 +266,31 @@ func resolveProjectConfigFileFromWorkingDir() *string {
 
 	for _, projectConfigName := range projectConfigNames {
 
-		path := workDir
+		p := workDir
 
-		_, err = os.Stat(filepath.Join(path, projectConfigName))
+		_, err = os.Stat(filepath.Join(p, projectConfigName))
 		found := true
 
 		for os.IsNotExist(err) {
 
 			// stop when leaving a git repo
-			if gitDir, statErr := os.Stat(filepath.Join(path, ".git")); statErr == nil && gitDir.IsDir() {
+			if gitDir, statErr := os.Stat(filepath.Join(p, ".git")); statErr == nil && gitDir.IsDir() {
 				found = false
 				break
 			}
 
-			oldPath := path
+			oldPath := p
 
-			if path = filepath.Dir(oldPath); path == oldPath {
+			if p = filepath.Dir(oldPath); p == oldPath {
 				output.Debugf("cannot find project config file: %s", projectConfigName)
 				found = false
 				break
 			}
-			_, err = os.Stat(filepath.Join(path, projectConfigName))
+			_, err = os.Stat(filepath.Join(p, projectConfigName))
 		}
 
 		if found {
-			configFile := filepath.Join(path, projectConfigName)
+			configFile := filepath.Join(p, projectConfigName)
 			return &configFile
 		}
 	}
@@ -239,42 +318,67 @@ func mapEnvVariables() {
 }
 
 // generateDefaultConfig generates default config in case there is no config
-func generateDefaultConfig() error {
+func generateDefaultConfig(name string, viperInstance *viper.Viper) error {
 
-	cfgFile := filepath.Join(os.ExpandEnv(configPaths[0]), "config.yml")
+	filename := name + ".yml"
+	cfgFile := filepath.Join(os.ExpandEnv(configPaths[0]), filename)
 
-	if os.Getenv("KAFKA_CTL_CONFIG") != "" {
+	if name == configFileName && os.Getenv(ConfigEnvVariable) != "" {
 		// use config file provided via env
-		cfgFile = os.Getenv("KAFKA_CTL_CONFIG")
-	} else if os.Getenv("SNAP_REAL_HOME") != "" {
-		// use different configFile when running in snap
-		for _, configPath := range configPaths {
-			if strings.Contains(configPath, "$SNAP_REAL_HOME") {
-				cfgFile = filepath.Join(os.ExpandEnv(configPath), "config.yml")
-				break
-			}
-		}
+		cfgFile = os.Getenv(ConfigEnvVariable)
+	} else if name == writableConfigFileName && os.Getenv(WritableConfigEnvVariable) != "" {
+		// use config file provided via env
+		cfgFile = os.Getenv(WritableConfigEnvVariable)
+	} else if name == writableConfigFileName && os.Getenv(ConfigEnvVariable) != "" {
+		// use config file provided via env as base path
+		cfgFile = filepath.Join(path.Dir(os.Getenv(ConfigEnvVariable)), filename)
 	} else if runtime.GOOS == "windows" {
 		// use different configFile when running on windows
 		for _, configPath := range configPaths {
 			if strings.Contains(configPath, "$APPDATA") {
-				cfgFile = filepath.Join(os.ExpandEnv(configPath), "config.yml")
+				cfgFile = filepath.Join(os.ExpandEnv(configPath), filename)
 				break
 			}
 		}
 	}
 
-	if err := os.MkdirAll(filepath.Dir(cfgFile), os.FileMode(0700)); err != nil {
-		return err
+	if name == configFileName {
+		viperInstance.SetDefault("contexts.default.brokers", []string{"localhost:9092"})
+	} else {
+		viperInstance.SetDefault("current-context", getInitialCurrentContext(cfgFile))
 	}
 
-	viper.SetDefault("contexts.default.brokers", []string{"localhost:9092"})
-	viper.SetDefault("current-context", "default")
+	if err := os.MkdirAll(filepath.Dir(cfgFile), os.FileMode(0700)); err != nil {
+		output.Warnf("cannot creating config file directory: %v", err)
+		return nil
+	}
 
-	if err := viper.WriteConfigAs(cfgFile); err != nil {
-		return err
+	if err := viperInstance.WriteConfigAs(cfgFile); err != nil {
+		output.Warnf("cannot write config file=%s: %v", cfgFile, err)
+		return nil
 	}
 
 	output.Debugf("generated default config at %s", cfgFile)
+
+	// We read generated config now
+	if err := loadConfig(name, viperInstance, &cfgFile); err != nil {
+		return fmt.Errorf("error reading config file: %s (%v)", viperInstance.ConfigFileUsed(), err.Error())
+	}
+
 	return nil
+}
+
+func getInitialCurrentContext(writableConfigFile string) string {
+
+	if viper.GetString("current-context") != "" {
+		output.Warnf(fmt.Sprintf(`the configuration of the current context is now managed in a separate file: %s
+the parameter “current-context” in config.yml is no longer used. Please remove it to avoid confusion.`, writableConfigFile))
+		return viper.GetString("current-context")
+	}
+
+	availableContexts := ListAvailableContexts()
+	if len(availableContexts) > 0 {
+		return availableContexts[0]
+	}
+	return ""
 }
