@@ -3,63 +3,50 @@ package producer
 import (
 	"encoding/binary"
 
-	"github.com/deviceinsight/kafkactl/v5/internal/helpers/avro"
-
 	"github.com/IBM/sarama"
+	"github.com/deviceinsight/kafkactl/v5/internal"
+	"github.com/deviceinsight/kafkactl/v5/internal/helpers/avro"
+	"github.com/riferrei/srclient"
+
 	"github.com/deviceinsight/kafkactl/v5/internal/util"
-	schemaregistry "github.com/landoop/schema-registry"
 	"github.com/linkedin/goavro/v2"
 	"github.com/pkg/errors"
 )
 
 type AvroMessageSerializer struct {
-	topic              string
-	avroSchemaRegistry string
-	jsonCodec          avro.JSONCodec
-	client             *schemaregistry.Client
+	topic     string
+	jsonCodec avro.JSONCodec
+	client    *internal.CachingSchemaRegistry
 }
 
-func CreateAvroMessageSerializer(topic string, avroSchemaRegistry string, jsonCodec avro.JSONCodec) (AvroMessageSerializer, error) {
+func (serializer AvroMessageSerializer) CanSerializeValue(topic string) (bool, error) {
+	return serializer.client.SubjectOfTypeExists(topic+"-value", srclient.Avro)
+}
 
+func (serializer AvroMessageSerializer) CanSerializeKey(topic string) (bool, error) {
+	return serializer.client.SubjectOfTypeExists(topic+"-key", srclient.Avro)
+}
+
+func (serializer AvroMessageSerializer) SerializeValue(value []byte, flags Flags) ([]byte, error) {
+	return serializer.encode(value, flags.ValueSchemaVersion, serializer.topic+"-value")
+}
+
+func (serializer AvroMessageSerializer) SerializeKey(key []byte, flags Flags) ([]byte, error) {
+	return serializer.encode(key, flags.KeySchemaVersion, serializer.topic+"-key")
+}
+
+func (serializer AvroMessageSerializer) encode(rawData []byte, schemaVersion int, subject string) ([]byte, error) {
+
+	var schema *srclient.Schema
 	var err error
-
-	serializer := AvroMessageSerializer{topic: topic, avroSchemaRegistry: avroSchemaRegistry, jsonCodec: jsonCodec}
-
-	serializer.client, err = schemaregistry.NewClient(serializer.avroSchemaRegistry)
-
-	if err != nil {
-		return serializer, errors.Wrap(err, "failed to create schema registry client: ")
-	}
-
-	return serializer, nil
-}
-
-func (serializer AvroMessageSerializer) encode(rawData []byte, schemaVersion int, avroSchemaType string) ([]byte, error) {
-
-	subject := serializer.topic + "-" + avroSchemaType
-
-	subjects, err := serializer.client.Subjects()
-
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to list available avro schemas")
-	}
-
-	if !util.ContainsString(subjects, subject) {
-		// does not seem to be avro data
-		return rawData, nil
-	}
-
-	var schema schemaregistry.Schema
 
 	if schemaVersion == -1 {
 		schema, err = serializer.client.GetLatestSchema(subject)
-
 		if err != nil {
 			return nil, errors.Errorf("failed to find latest avro schema for subject: %s (%v)", subject, err)
 		}
 	} else {
-		schema, err = serializer.client.GetSchemaBySubject(subject, schemaVersion)
-
+		schema, err = serializer.client.GetSchemaByVersion(subject, schemaVersion)
 		if err != nil {
 			return nil, errors.Errorf("failed to find avro schema for subject: %s id: %d (%v)", subject, schemaVersion, err)
 		}
@@ -68,9 +55,9 @@ func (serializer AvroMessageSerializer) encode(rawData []byte, schemaVersion int
 	var avroCodec *goavro.Codec
 
 	if serializer.jsonCodec == avro.Avro {
-		avroCodec, err = goavro.NewCodec(schema.Schema)
+		avroCodec, err = goavro.NewCodec(schema.Schema())
 	} else {
-		avroCodec, err = goavro.NewCodecForStandardJSONFull(schema.Schema)
+		avroCodec, err = goavro.NewCodecForStandardJSONFull(schema.Schema())
 	}
 
 	if err != nil {
@@ -89,19 +76,19 @@ func (serializer AvroMessageSerializer) encode(rawData []byte, schemaVersion int
 
 	// https://docs.confluent.io/current/schema-registry/docs/serializer-formatter.html#wire-format
 	versionBytes := make([]byte, 5)
-	binary.BigEndian.PutUint32(versionBytes[1:], uint32(schema.ID))
+	binary.BigEndian.PutUint32(versionBytes[1:], uint32(schema.ID()))
 
 	return append(versionBytes, data...), nil
 }
 
 func (serializer AvroMessageSerializer) CanSerialize(topic string) (bool, error) {
-
-	subjects, err := serializer.client.Subjects()
-
+	subjects, err := serializer.client.GetSubjects()
 	if err != nil {
 		return false, errors.Wrap(err, "failed to list available avro schemas")
 	}
 
+	// we currently only implement the TopicNameStrategy to map from a message to the corresponding schema
+	// in this strategy the name of the schema subject is derived from topic name i.e. `topic+"-key"`, `topic+"-value"`
 	if util.ContainsString(subjects, topic+"-key") {
 		return true, nil
 	} else if util.ContainsString(subjects, topic+"-value") {
@@ -111,7 +98,6 @@ func (serializer AvroMessageSerializer) CanSerialize(topic string) (bool, error)
 }
 
 func (serializer AvroMessageSerializer) Serialize(key, value []byte, flags Flags) (*sarama.ProducerMessage, error) {
-
 	recordHeaders, err := createRecordHeaders(flags)
 	if err != nil {
 		return nil, err
