@@ -9,9 +9,11 @@ import (
 	"github.com/pkg/errors"
 )
 
-type MessageSerializer interface {
-	CanSerialize(topic string) (bool, error)
-	Serialize(key, value []byte, flags Flags) (*sarama.ProducerMessage, error)
+type messageSerializer interface {
+	CanSerializeValue(topic string) (bool, error)
+	CanSerializeKey(topic string) (bool, error)
+	SerializeValue(value []byte, flags Flags) ([]byte, error)
+	SerializeKey(key []byte, flags Flags) ([]byte, error)
 }
 
 const (
@@ -20,7 +22,6 @@ const (
 )
 
 func parseHeader(raw string) (key, value string, err error) {
-
 	// use a regex to split in order to handle escaped colons
 	re := regexp.MustCompile(`[^\\]:`)
 
@@ -63,27 +64,19 @@ func decodeBytes(data []byte, encoding string) ([]byte, error) {
 	switch encoding {
 	case HEX:
 		out = make([]byte, hex.DecodedLen(len(data)))
-		if _, err := hex.Decode(out, data); err != nil {
+		bytelen, err := hex.Decode(out, data)
+		if err != nil {
 			return nil, err
 		}
-		return out, nil
+		return out[:bytelen], nil
 	case BASE64:
 		out = make([]byte, base64.StdEncoding.DecodedLen(len(data)))
-		if _, err := base64.StdEncoding.Decode(out, data); err != nil {
+		bytelen, err := base64.StdEncoding.Decode(out, data)
+		if err != nil {
 			return nil, err
 		}
-		return out[:clen(out)], nil
+		return out[:bytelen], nil
 	default:
 		return data, nil
 	}
-}
-
-// https://stackoverflow.com/a/27834860/12143351
-func clen(n []byte) int {
-	for i := 0; i < len(n); i++ {
-		if n[i] == 0 {
-			return i
-		}
-	}
-	return len(n)
 }
