@@ -1,8 +1,9 @@
 package input
 
 type Message struct {
-	Key   string `json:"key"`
-	Value string `json:"value"`
+	Key     *string           `json:"key"`
+	Value   *string           `json:"value"`
+	Headers map[string]string `json:"headers,omitempty"`
 }
 
 type Parser interface {
