@@ -9,15 +9,16 @@ import (
 	"testing"
 	"time"
 
+	"github.com/deviceinsight/kafkactl/v5/internal/output"
+
 	"github.com/Rican7/retry"
 	"github.com/Rican7/retry/backoff"
 	"github.com/Rican7/retry/strategy"
 	"github.com/deviceinsight/kafkactl/v5/internal/util"
-	schemaregistry "github.com/landoop/schema-registry"
+	"github.com/riferrei/srclient"
 )
 
 func CreateTopic(t *testing.T, topicPrefix string, flags ...string) string {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 	topicName := GetPrefixedName(topicPrefix)
 
@@ -34,25 +35,36 @@ func CreateTopic(t *testing.T, topicPrefix string, flags ...string) string {
 	return topicName
 }
 
-func CreateAvroTopic(t *testing.T, topicPrefix, keySchema, valueSchema string, flags ...string) string {
+func RegisterSchema(t *testing.T, subjectName, schema string, schemaType srclient.SchemaType, references ...srclient.Reference) {
+	schemaRegistry := srclient.NewSchemaRegistryClient("http://localhost:18081")
 
+	if schema, err := schemaRegistry.CreateSchema(subjectName, schema, schemaType, references...); err != nil {
+		t.Fatalf("unable to register schema for value: %v", err)
+	} else {
+		output.TestLogf("registered schema %q with ID=%d", schema, schema.ID())
+	}
+}
+
+func CreateTopicWithSchema(t *testing.T, topicPrefix, keySchema, valueSchema string, schemaType srclient.SchemaType,
+	flags ...string,
+) string {
 	topicName := CreateTopic(t, topicPrefix, flags...)
 
-	schemaRegistry, err := schemaregistry.NewClient("localhost:18081")
-
-	if err != nil {
-		t.Fatalf("failed to create schema registry client: %v", err)
-	}
+	schemaRegistry := srclient.NewSchemaRegistryClient("http://localhost:18081")
 
 	if keySchema != "" {
-		if _, err := schemaRegistry.RegisterNewSchema(topicName+"-key", keySchema); err != nil {
+		if schema, err := schemaRegistry.CreateSchema(topicName+"-key", keySchema, schemaType); err != nil {
 			t.Fatalf("unable to register schema for key: %v", err)
+		} else {
+			output.TestLogf("registered schema %q with ID=%d", topicName+"-key", schema.ID())
 		}
 	}
 
 	if valueSchema != "" {
-		if _, err := schemaRegistry.RegisterNewSchema(topicName+"-value", valueSchema); err != nil {
+		if schema, err := schemaRegistry.CreateSchema(topicName+"-value", valueSchema, schemaType); err != nil {
 			t.Fatalf("unable to register schema for value: %v", err)
+		} else {
+			output.TestLogf("registered schema %q with ID=%d", topicName+"-value", schema.ID())
 		}
 	}
 
@@ -60,12 +72,10 @@ func CreateAvroTopic(t *testing.T, topicPrefix, keySchema, valueSchema string, f
 }
 
 func VerifyTopicExists(t *testing.T, topic string) {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 
 	findTopic := func(_ uint) error {
 		_, err := kafkaCtl.Execute("get", "topics", "-o", "compact")
-
 		if err != nil {
 			return err
 		}
@@ -81,7 +91,6 @@ func VerifyTopicExists(t *testing.T, topic string) {
 		strategy.Limit(5),
 		strategy.Backoff(backoff.Linear(10*time.Millisecond)),
 	)
-
 	if err != nil {
 		t.Fatalf("could not find topic %s: %v", topic, err)
 	}
@@ -92,7 +101,6 @@ func VerifyTopicExists(t *testing.T, topic string) {
 }
 
 func CreateConsumerGroup(t *testing.T, groupPrefix string, topics ...string) string {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 	groupName := GetPrefixedName(groupPrefix)
 
@@ -114,7 +122,6 @@ func CreateConsumerGroup(t *testing.T, groupPrefix string, topics ...string) str
 }
 
 func ProduceMessage(t *testing.T, topic, key, value string, expectedPartition, expectedOffset int64) {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 
 	if _, err := kafkaCtl.Execute("produce", topic, "--key", key, "--value", value); err != nil {
@@ -125,7 +132,6 @@ func ProduceMessage(t *testing.T, topic, key, value string, expectedPartition, e
 }
 
 func ProduceMessageOnPartition(t *testing.T, topic, key, value string, partition int32, expectedOffset int64) {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 
 	if _, err := kafkaCtl.Execute("produce", topic, "--key", key, "--value", value, "--partition", strconv.FormatInt(int64(partition), 10)); err != nil {
@@ -136,12 +142,10 @@ func ProduceMessageOnPartition(t *testing.T, topic, key, value string, partition
 }
 
 func VerifyGroupExists(t *testing.T, group string) {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 
 	findConsumerGroup := func(_ uint) error {
 		_, err := kafkaCtl.Execute("get", "cg", "-o", "compact")
-
 		if err != nil {
 			return err
 		}
@@ -157,7 +161,6 @@ func VerifyGroupExists(t *testing.T, group string) {
 		strategy.Limit(5),
 		strategy.Backoff(backoff.Linear(10*time.Millisecond)),
 	)
-
 	if err != nil {
 		t.Fatalf("could not find group %s: %v", group, err)
 	}
@@ -168,14 +171,12 @@ func VerifyGroupExists(t *testing.T, group string) {
 }
 
 func VerifyConsumerGroupOffset(t *testing.T, group, topic string, expectedConsumerOffset int) {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 
 	consumerOffsetRegex, _ := regexp.Compile(`consumerOffset: (\d)`)
 
 	verifyConsumerOffset := func(_ uint) error {
 		_, err := kafkaCtl.Execute("describe", "cg", group, "--topic", topic, "-o", "yaml")
-
 		if err != nil {
 			return err
 		}
@@ -197,21 +198,18 @@ func VerifyConsumerGroupOffset(t *testing.T, group, topic string, expectedConsum
 		strategy.Limit(5),
 		strategy.Backoff(backoff.Linear(10*time.Millisecond)),
 	)
-
 	if err != nil {
 		t.Fatalf("failed to verify offset for group=%s topic=%s: %v", group, topic, err)
 	}
 }
 
 func VerifyTopicNotInConsumerGroup(t *testing.T, group, topic string) {
-
 	kafkaCtl := CreateKafkaCtlCommand()
 
 	emptyTopicsRegex, _ := regexp.Compile(`topics: \[]`)
 
 	verifyTopicNotInGroup := func(_ uint) error {
 		_, err := kafkaCtl.Execute("describe", "cg", group, "--topic", topic, "-o", "yaml")
-
 		if err != nil {
 			return err
 		}
@@ -227,7 +225,6 @@ func VerifyTopicNotInConsumerGroup(t *testing.T, group, topic string) {
 		strategy.Limit(5),
 		strategy.Backoff(backoff.Linear(10*time.Millisecond)),
 	)
-
 	if err != nil {
 		t.Fatalf("failed to verify topic=%s not in group=%s: %v", topic, group, err)
 	}
